### 2024-12-30 #7620 @editorialbot
[PRE REVIEW]: Mesa 3: Agent-based modelling with Python in 2024
**Submitting author:** @EwoutH (<a href="http://orcid.org/0009-0002-0805-3425">Ewout ter Hoeven</a>)
**Repository:** https://github.com/projectmesa/mesa
**Branch with paper.md** (empty if default branch): paper_2024
**Version:** v3.1
