### 2024-12-30 #7620 @danielskatz
👋 @EwoutH - thanks for your submission. Here are some items that need work.

1. Please add countries to all affiliations in your .md file.
2. Note that when you refer explicitly to a reference, as in the second reference in your paper, you shouldn't use []s around it in the .md file, as shown in the JOSS [example paper](https://joss.readthedocs.io/en/latest/example_paper.html).
3. I see that your paper is a bit longer than normal for JOSS (1500 words vs a suggested max of 1000). Please consider if there is anything you could move to the repo or documentation, and point to in the paper.
4. You could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect.

Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when you've made changes in the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Finally, can we close https://github.com/openjournals/joss/issues/1403 given that you have now submitted this paper, which seems fine to proceed with, other than the minor points above?
