### 2024-12-30 6e42b29
paper: Add DOIs to references
Add DOIs (Digital Object Identifiers) for most papers (except the oldest, which don't have nicely formatted ones).

For some reason Google Scholar didn't do this by default, which is kind of weird.
