+++ paper/paper.bib
@@ -1,10 +1,10 @@
 @book{epstein_axtell_1996,
-author = {Joshua M. Epstein and Robert Axtell},
-title = {Growing Artificial Societies: Social Science from the Bottom Up},
-year = {1996},
-publisher = {Brookings Institution Press and MIT Press},
-address = {Washington, DC, and Cambridge, MA},
-isbn = {9780262550253},
+  author = {Joshua M. Epstein and Robert Axtell},
+  title = {Growing Artificial Societies: Social Science from the Bottom Up},
+  year = {1996},
+  publisher = {Brookings Institution Press and MIT Press},
+  address = {Washington, DC, and Cambridge, MA},
+  isbn = {9780262550253},
 }
 
 @article{epstein1999,
@@ -19,25 +19,27 @@ isbn = {9780262550253},
 }
 
 @inproceedings{Dra_gulescu_2003,
-title={Statistical Mechanics of Money, Income, and Wealth: A Short Survey},
-volume={661},
-ISSN={0094-243X},
-url={http://dx.doi.org/10.1063/1.1571309},
-DOI={10.1063/1.1571309},
-booktitle={AIP Conference Proceedings},
-publisher={AIP},
-author={Drǎgulescu, Adrian A.},
-year={2003},
-pages={180–183} }
+  title={Statistical Mechanics of Money, Income, and Wealth: A Short Survey},
+  volume={661},
+  ISSN={0094-243X},
+  url={http://dx.doi.org/10.1063/1.1571309},
+  DOI={10.1063/1.1571309},
+  booktitle={AIP Conference Proceedings},
+  publisher={AIP},
+  author={Drǎgulescu, Adrian A.},
+  year={2003},
+  pages={180–183}
+}
 
 @book{schelling_2006,
-author = {Thomas C. Schelling},
-title = {Micromotives and Macrobehavior},
-year = {2006},
-publisher = {W. W. Norton & Company},
-address = {New York},
-edition = {Revised},
-isbn = {9780393329469},
+  author = {Thomas C. Schelling},
+  title = {Micromotives and Macrobehavior},
+  year = {2006},
+  publisher = {W. W. Norton & Company},
+  address = {New York},
+  edition = {Revised},
+  isbn = {9780393329469},
+  doi = {10.2307/134812}
 }
 
 @inproceedings{masad2015,
@@ -46,16 +48,18 @@ isbn = {9780393329469},
   booktitle={SciPy},
   pages={51--58},
   year={2015},
-  organization={Citeseer}
+  organization={Citeseer},
+  doi={10.25080/Majora-7b98e3ed-009}
 }
 
 @inproceedings{kazil2020,
   title={Utilizing python for agent-based modeling: The mesa framework},
   author={Kazil, Jackie and Masad, David and Crooks, Andrew},
-  booktitle={Social, Cultural, and Behavioral Modeling: 13th International Conference, SBP-BRiMS 2020, Washington, DC, USA, October 18--21, 2020, Proceedings 13},
+  booktitle={Social, Cultural, and Behavioral Modeling: 13th International Conference, SBP-BRiMS 2020},
   pages={308--317},
   year={2020},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-61255-9_30}
 }
 
 @inproceedings{wang2022mesa,
@@ -63,7 +67,8 @@ isbn = {9780393329469},
   author={Wang, Boyu and Hess, Vincent and Crooks, Andrew},
[…]
