### 2024-12-30 5975ca5
Fix citation format in JOSS paper
