### 2024-12-30 #7620 @EwoutH
@danielskatz Thanks for the very quick pre-review assignment!

1. Affiliations: would this be formatted correctly?
    ```md
    affiliations:
      - name: Delft University of Technology (Faculty of Technology, Policy and Management), The Netherlands
        index: 1
      - name: Independent Researcher, Germany
        index: 2
      - name: George Mason University (Department of Computational Social Science), USA
        index: 3
      - name: University at Buffalo (Department of Geography), USA
        index: 4
    ```
2. Explicit citations: Good catch, should be addressed in https://github.com/projectmesa/mesa/commit/a044848b428c114797a7286ed84716e6369284bd
3. Length: It's indeed a bit on the longer end, but we hope it's okay. It's 5 years of development by a large team of a major package in our space, and we went over it diligently multiple times to shorten it where feasible. We also hope the code blocks can be seem as "illustrations".
4. DOIs: You're right, for some reason Google Scholar didn't included them when generating BibTex. I've included them in https://github.com/projectmesa/mesa/commit/6e42b29ab9fdfb4845d96f28fcf88beeefe86e26.
