### 2024-12-30 3f65fc6
paper: Move rht from author to acknowledgements
Since we haven't heard anything from rht, I don't think we can add him as an author without his consent. Like we discussed, we recognize him in the Acknowledgements instead.
