### 2024-12-30 eb37092
paper: End with references header
The references get added automatically but a header does not. This fixed that.
