### 2024-12-30 #7620 @EwoutH
@danielskatz I should now have fully addressed all comments, thanks for your swift replies! If you need anything else from our side, please let us know.

> @EwoutH if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

No one specific, but there seems quite a list of JOSS reviewers experienced with ABM: [Programming language: Python, Keyword: agent-based](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=&keywords=agent-based&name=&commit=Search)
