### 2024-12-30 2b4b8ac
paper: Fix lists and code blocks jump in concistently
- Fix lists by starting them always after a new line
- Jump in the codeblocks to make them better readable
