+++ paper/paper.md
@@ -74,14 +74,14 @@ Central to ABMs are the autonomous heterogeneous agents. Mesa provides a variety
 One significant advancement of Mesa 3+ is expanded functionality around agent management. The new [`AgentSet`](https://mesa.readthedocs.io/latest/apis/agent.html#mesa.agent.AgentSet) class provides methods that allow users to filter, group, and analyze agents, making it easier to express complex model logic.
 
 ```python
-# Select wealthy agents and calculate average wealth
-wealthy = model.agents.select(lambda a: a.wealth > 1000)
-avg_wealth = wealthy.agg("wealth", func=np.mean)
-
-# Group agents by type and apply behaviors
-grouped = model.agents.groupby("species")
-for species, agents in grouped:
-    agents.shuffle_do("reproduce")
+    # Select wealthy agents and calculate average wealth
+    wealthy = model.agents.select(lambda a: a.wealth > 1000)
+    avg_wealth = wealthy.agg("wealth", func=np.mean)
+    
+    # Group agents by type and apply behaviors
+    grouped = model.agents.groupby("species")
+    for species, agents in grouped:
+        agents.shuffle_do("reproduce")
 ```
 
 ### Spaces
@@ -92,10 +92,11 @@ Mesa 3 provides both discrete (cell-based) and continuous space implementations.
 - Voronoi-based: `VoronoiMesh` for irregular tessellations
 
 Example grid creation:
+
 ```python
-grid = OrthogonalVonNeumannGrid(
-    (width, height), torus=False, random=model.random
-)
+    grid = OrthogonalVonNeumannGrid(
+        (width, height), torus=False, random=model.random
+    )
 ```
 
 Mesa provides specialized agent classes for spatial interactions in the discrete spaces:
@@ -105,74 +106,77 @@ Mesa provides specialized agent classes for spatial interactions in the discrete
 - `Grid2DMovingAgent`: Extends `CellAgent` with directional movement methods
 
 All discrete spaces support PropertyLayers - efficient numpy-based arrays for storing cell-level properties:
+
 ```python
-grid.create_property_layer("elevation", default_value=10)
-high_ground = grid.elevation.select_cells(lambda x: x > 50)
+    grid.create_property_layer("elevation", default_value=10)
+    high_ground = grid.elevation.select_cells(lambda x: x > 50)
 ```
 
 For models where agents need to move continuously through space rather than between discrete locations, `ContinuousSpace` allows agents to occupy any coordinate within defined boundaries:
+
 ```python
-space = ContinuousSpace(x_max, y_max, torus=True)
-space.move_agent(agent, (new_x, new_y))
+    space = ContinuousSpace(x_max, y_max, torus=True)
+    space.move_agent(agent, (new_x, new_y))
 ```
 
 ### Time advancement
 Typically, ABMs rely on incremental time progression or ticks. For each tick, the step method of the model is called. This activates the agents in some way. The most frequently implemented approach is shown below, which runs a model for 100 ticks.
 
 ```python
-model = Model(seed=42)
-
-for _ in range(100):
-    model.step()
+    model = Model(seed=42)
+    
+    for _ in range(100):
+        model.step()
 ```
 
 Generally, within the step method of the model, one activates all the agents. The newly added `AgentSet` class provides a more flexible way to activate agents replacing the fixed patterns previously available.
 
 ```python
-model.agents.do("step")  # Sequential activation
-
-model.agents.shuffle_do("step")  # Random activation
-
-# Multi-stage activation:
-for stage in ["move", "eat", "reproduce"]:
-    model.agents.do(stage)
-
-# Activation by agent subclass:
-for klass in model.agent_types:
[…]
