### 2024-12-30 11d3833
paper: Format a codeblock to a single line
