### 2024-12-30 6ef0b13
Update continuous_space.py
