### 2024-12-30 6a743f6
paper: Add rht as author
Add rht back as author, since he got back to us and indicated he wanted to be on the paper.
