### 2024-12-30 2782736
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
