### 2024-12-30 #7620 @EwoutH
We have one maybe unusual request: On of our authors (@rht) would like to go under that name, his pseudonym he did all his work on Mesa under over the past years. He has [an ORCID](https://orcid.org/0009-0002-6902-111X), but would like to omit his full name and country. Would that be possible in JOSS?
