### 2024-12-30 d0b05ec
Update continuous_space.py
