### 2024-12-31 c7c8311
refactor meta-agent
- allow for deliberate meta-agents creation
- allow for combinatorics
- allow for dynamic agent creation

fix methods-functions; add tests
