+++ mesa/examples/basic/alliance_formation_model/model.py
@@ -3,10 +3,25 @@ import numpy as np
 
 import mesa
 from mesa.examples.basic.alliance_formation_model.agents import AllianceAgent
+from mesa.experimental.meta_agents.meta_agent import find_combinations
+from mesa.experimental.meta_agents.multi_levels import multi_level_agents
 
 
-class AllianceModel(mesa.Model):
+class MultiLevelAllianceModel(mesa.Model):
+    """
+    Model for simulating multi-level alliances among agents.
+    """
+
     def __init__(self, n=50, mean=0.5, std_dev=0.1, seed=42):
+        """
+        Initialize the model.
+
+        Args:
+            n (int): Number of agents.
+            mean (float): Mean value for normal distribution.
+            std_dev (float): Standard deviation for normal distribution.
+            seed (int): Random seed.
+        """
         super().__init__(seed=seed)
         self.population = n
         self.network = nx.Graph()  # Initialize the network
@@ -19,21 +34,147 @@ class AllianceModel(mesa.Model):
         position = np.clip(position, 0, 1)
         AllianceAgent.create_agents(self, n, power, position)
         agent_ids = [
-            (agent.unique_id, {"size": 300, "hierarchy": 0}) for agent in self.agents
+            (agent.unique_id, {"size": 300, "level": 0}) for agent in self.agents
         ]
         self.network.add_nodes_from(agent_ids)
 
     def add_link(self, meta_agent, agents):
+        """
+        Add links between a meta agent and its constituent agents in the network.
+
+        Args:
+            meta_agent (MetaAgent): The meta agent.
+            agents (list): List of agents.
+        """
         for agent in agents:
             self.network.add_edge(meta_agent.unique_id, agent.unique_id)
 
-    def step(self):
-        for agent_class in list(
-            self.agent_types
-        ):  # Convert to list to avoid modification during iteration
-            self.agents_by_type[agent_class].shuffle_do("form_alliance")
+    def calculate_shapley_value(self, agents):
+        """
+        Calculate the Shapley value of the two agents.
 
-            # Update graph
-            if agent_class is not AllianceAgent:
-                for meta_agent in self.agents_by_type[agent_class]:
-                    self.add_link(meta_agent, meta_agent.agents)
+        Args:
+            agents (list): List of agents.
+
+        Returns:
+            tuple: Potential utility, new position, and level.
+        """
+        positions = agents.get("position")
+        new_position = 1 - (max(positions) - min(positions))
+        potential_utility = agents.agg("power", sum) * 1.2 * new_position
+
+        value_0 = 0.5 * agents[0].power + 0.5 * (potential_utility - agents[1].power)
+        value_1 = 0.5 * agents[1].power + 0.5 * (potential_utility - agents[0].power)
+
+        if value_0 > agents[0].power and value_1 > agents[1].power:
+            if agents[0].level > agents[1].level:
+                level = agents[0].level
+            elif agents[0].level == agents[1].level:
+                level = agents[0].level + 1
+            else:
+                level = agents[1].level
+
+            return potential_utility, new_position, level
+
+    def only_best_combination(self, combinations):
+        """
+        Filter to keep only the best combination for each agent.
+
+        Args:
+            combinations (list): List of combinations.
+
+        Returns:
+            dict: Unique combinations.
+        """
+        best = {}
+        # Determine best option for EACH agent
+        for group, value in combinations:
+            agent_ids = sorted(group.get("unique_id"))  # by default is bilateral
+            # Deal with all possibilities
+            if (
+                agent_ids[0] not in best and agent_ids[1] not in best
+            ):  # if neither in add both
+                best[agent_ids[0]] = [group, value, agent_ids]
[…]
