+++ tests/test_meta_agents.py
@@ -3,7 +3,25 @@
 import pytest
 
 from mesa import Agent, Model
-from mesa.experimental.meta_agents.meta_agents import create_meta_agent
+from mesa.experimental.meta_agents.meta_agent import (
+    MetaAgent,
+    evaluate_combination,
+    find_combinations,
+)
+from mesa.experimental.meta_agents.multi_levels import multi_level_agents
+
+
+class CustomAgent(Agent):
+    """A custom agent with additional attributes and methods."""
+
+    def __init__(self, model):
+        """A custom agent constructor."""
+        super().__init__(model)
+        self.custom_attribute = "custom_value"
+
+    def custom_method(self):
+        """A custom agent method."""
+        return "custom_method_value"
 
 
 @pytest.fixture
@@ -17,8 +35,7 @@ def setup_agents():
     agent1 = Agent(model)
     agent2 = Agent(model)
     agent3 = Agent(model)
-    agent4 = Agent(model)
-    agent4.custom_attribute = "custom_value"
+    agent4 = CustomAgent(model)
     agents = [agent1, agent2, agent3, agent4]
     return model, agents
 
@@ -30,20 +47,23 @@ def test_create_meta_agent_new_class(setup_agents):
         setup_agents (tuple): The model and agents fixture.
     """
     model, agents = setup_agents
-    meta_agent = create_meta_agent(
+    meta_agent = multi_level_agents(
         model,
         "MetaAgentClass",
         agents,
         meta_attributes={"attribute1": "value1"},
-        meta_functions={"function1": lambda self: "function1"},
+        meta_methods={"function1": lambda self: "function1"},
         retain_subagent_attributes=True,
+        retain_subagent_methods=True,
     )
     assert meta_agent is not None
     assert meta_agent.attribute1 == "value1"
     assert meta_agent.function1() == "function1"
-    assert meta_agent.agents == set(agents)
+    assert meta_agent._subset == set(agents)
     assert hasattr(meta_agent, "custom_attribute")
     assert meta_agent.custom_attribute == "custom_value"
+    assert hasattr(meta_agent, "custom_method")
+    assert meta_agent.custom_method() == "custom_method_value"
 
 
 def test_create_meta_agent_existing_class(setup_agents):
@@ -54,32 +74,35 @@ def test_create_meta_agent_existing_class(setup_agents):
     """
     model, agents = setup_agents
 
-    # Create Met Agent Class
-    meta_agent = create_meta_agent(
+    # Create Meta Agent Class
+    meta_agent = multi_level_agents(
         model,
         "MetaAgentClass",
         [agents[0], agents[2]],
         meta_attributes={"attribute1": "value1"},
-        meta_functions={"function1": lambda self: "function1"},
+        meta_methods={"function1": lambda self: "function1"},
     )
 
     # Create new meta-agent instance with existing class
-    meta_agent2 = create_meta_agent(
+    meta_agent2 = multi_level_agents(
         model,
         "MetaAgentClass",
         [agents[1], agents[3]],
         meta_attributes={"attribute2": "value2"},
-        meta_functions={"function2": lambda self: "function2"},
+        meta_methods={"function2": lambda self: "function2"},
         retain_subagent_attributes=True,
+        retain_subagent_methods=True,
     )
     assert meta_agent is not None
     assert meta_agent2.attribute2 == "value2"
     assert meta_agent.function1() == "function1"
-    assert meta_agent.agents == {agents[2], agents[0]}
+    assert meta_agent._subset == {agents[2], agents[0]}
     assert meta_agent2.function2() == "function2"
-    assert meta_agent2.agents == {agents[1], agents[3]}
[…]
