+++ mesa/examples/basic/alliance_formation_model/app.py
@@ -3,12 +3,9 @@ import networkx as nx
 import solara
 from matplotlib.figure import Figure
 
-from mesa.examples.basic.alliance_formation_model.model import AllianceModel
-from mesa.mesa_logging import DEBUG, log_to_stderr
+from mesa.examples.basic.alliance_formation_model.model import MultiLevelAllianceModel
 from mesa.visualization import SolaraViz
 from mesa.visualization.utils import update_counter
-
-log_to_stderr(DEBUG)
 
 model_params = {
     "seed": {
@@ -37,7 +34,7 @@ model_params = {
 def plot_network(model):
     update_counter.get()
     g = model.network
-    pos = nx.kamada_kawai_layout(g)
+    pos = nx.fruchterman_reingold_layout(g)
     fig = Figure()
     ax = fig.subplots()
     labels = {agent.unique_id: agent.unique_id for agent in model.agents}
@@ -58,13 +55,14 @@ def plot_network(model):
 
 
 # Create initial model instance
-model = AllianceModel(50)
+model = MultiLevelAllianceModel(50)
 
 # Create the SolaraViz page. This will automatically create a server and display the
 # visualization elements in a web browser.
 # Display it using the following command in the example directory:
 # solara run app.py
 # It will automatically update and display any changes made to this file
+
 page = SolaraViz(
     model,
     components=[plot_network],
