+++ mesa/experimental/meta_agents/__init__.py
@@ -20,6 +20,7 @@ Goal is to assess usage and expand functionality.
 
 """
 
-from .meta_agents import create_meta_agent
+from .meta_agent import MetaAgent
+from .multi_levels import multi_level_agents
 
-__all__ = ["create_meta_agent"]
+__all__ = ["MetaAgent", "multi_level_agents"]
