+++ mesa/examples/__init__.py
@@ -1,3 +1,5 @@
+from multi_level_alliance.model import AllianceModel
+
 from mesa.examples.advanced.epstein_civil_violence.model import EpsteinCivilViolence
 from mesa.examples.advanced.pd_grid.model import PdGrid
 from mesa.examples.advanced.sugarscape_g1mt.model import SugarscapeG1mt
