+++ mesa/experimental/meta_agents/__init__.py
@@ -20,6 +20,6 @@ Goal is to assess usage and expand functionality.
 
 """
 
-from .meta_agents import create_meta_agent
+from .multi_levels import create_multi_levels
 
-__all__ = ["create_meta_agent"]
+__all__ = ["create_multi_levels"]

[…]
