+++ .vale.ini
@@ -1,4 +1,4 @@
-StylesPath = .vale/styles
+StylesPath = .github/styles
 MinAlertLevel = warning
 
 [*.md]
