### 2024-12-31 74d4971
added test file to check cli
