### 2024-12-31 201937e
Update .pre-commit-config.yaml
