+++ .pre-commit-config.yaml
@@ -3,6 +3,7 @@ ci:
 
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
+    # Ruff version.
     rev: v0.8.1
     hooks:
       # Run the linter.
