+++ mesa/experimental/continuous_space/__init__.py
@@ -1,4 +1,4 @@
-
-
 from mesa.experimental.continuous_space.continuous_space import ContinuousSpace
-from mesa.experimental.continuous_space.continuous_space_agents import ContinuousSpaceAgent
+from mesa.experimental.continuous_space.continuous_space_agents import (
+    ContinuousSpaceAgent,
+)
