### 2025-01-01 b6e351a
paper: Update title to 2025
Update the JOSS paper title to 2025. Also write modeling with on l (US spelling), to be consistent with the rest of the paper.
