### 2025-01-01 a66c54b
paper: Note experimental state of cell space and DEVS
As discussed in the tracking issue, this PR adds notes about the experimental state of both the cell space and DEVS.
