### 2025-01-01 955a2dc
paper: Fix whitespaces in codeblocks
