### 2025-01-01 85665c5
add scipy as dependency
