+++ mesa/experimental/continuous_space/continuous_space_agents.py
@@ -122,6 +122,10 @@ class ContinuousSpaceAgent(Agent):
         k += 1  # the distance calculation includes self, with a distance of 0, so we remove this later
         indices = np.argpartition(distances, k)[:k]
 
+        # fixme
+        #  indices mismatch, because it is the index in the distances
+        #  but this is not the same as the indices in agent_positions
+
         # don't forget to remove our self
         if include_distance:
             # don't forget to remove our self
