+++ mesa/experimental/continuous_space/continuous_space.py
@@ -104,7 +104,9 @@ class ContinuousSpace:
         """Calculate the distance between the point and all agents."""
         if self.torus:
             delta = np.abs(point[np.newaxis, :] - self._agent_positions)
-            delta = np.minimum(delta, 1 - delta)  # fixme, should be based on size or maxima?
+            delta = np.minimum(
+                delta, 1 - delta
+            )  # fixme, should be based on size or maxima?
             dists = np.linalg.norm(delta, axis=1)
         else:
             dists = cdist(point[np.newaxis, :], self.agent_positions)
