### 2025-01-01 a2d113e
ensure aligntment between agents and distances
