+++ mesa/experimental/continuous_space/continuous_space.py
@@ -40,7 +40,7 @@ class ContinuousSpace:
         self._agent_positions = np.zeros(
             (n_agents, self.dimensions.shape[0]), dtype=float
         )
-        self._agents = np.zeros((n_agents, ), dtype=object)
+        self._agents = np.zeros((n_agents,), dtype=object)
         self._positions_in_use = np.zeros(
             (n_agents,), dtype=bool
         )  # effectively a mask over _agent_positions
