### 2025-01-01 d8b62b0
Update continuous_space.py
