### 2025-01-01 3a985d8
tests for distance calculations
