### 2025-01-01 acd0ef0
ruff related fixes
