+++ mesa/experimental/continuous_space/__init__.py
@@ -2,3 +2,5 @@ from mesa.experimental.continuous_space.continuous_space import ContinuousSpace
 from mesa.experimental.continuous_space.continuous_space_agents import (
     ContinuousSpaceAgent,
 )
+
+__all__ = ["ContinuousSpace", "ContinuousSpaceAgent"]
