### 2025-01-02 4fd919f
Fix batch_run data collection on final step
