+++ mesa/examples/basic/boid_flockers/app.py
@@ -1,7 +1,7 @@
-import sys, os
+import os
+import sys
 
-
-sys.path.insert(0, os.path.abspath('../../../..'))
+sys.path.insert(0, os.path.abspath("../../../.."))
 
 from mesa.examples.basic.boid_flockers.model import BoidFlockers
 from mesa.visualization import Slider, SolaraViz, make_space_component
