### 2025-01-02 07e5da4
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
