+++ mesa/experimental/continuous_space/continuous_space.py
@@ -119,9 +119,7 @@ class ContinuousSpace:
 
         if self.torus:
             delta = np.abs(delta, out=delta)
-            delta = np.minimum(
-                delta, self.size - delta, out=delta
-            )
+            delta = np.minimum(delta, self.size - delta, out=delta)
 
         return delta
 
