### 2025-01-02 cc4a095
Update continuous_space.py
