### 2025-01-03 f90e9c9
paper: Update summary - wording.
