### 2025-01-03 0bc3938
paper: Update SoN - grammar, wording.
