### 2025-01-03 9970591
paper: Update language in CC.
