### 2025-01-03 f776adb
paper: Fix version number inconsistencies.
