+++ mesa/experimental/continuous_space/for_development.py
@@ -1,4 +1,4 @@
-# to be removed once further into the development
+"""to be removed once further into the development"""
 from random import Random
 
 import numpy as np
