### 2025-01-03 d6326fd
Update continuous_space.py
