+++ mesa/experimental/continuous_space/continuous_space.py
@@ -124,14 +124,14 @@ class ContinuousSpace:
         self._agents[index] = agent
         self._agent_to_index[agent] = index
         self._index_to_agent[index] = agent
-        self._update_stuff()
+        self.active_agents = self._agents[self._positions_in_use]
 
         return index
 
-    def _update_stuff(self):
-        """Fixme: Needs a better name."""
-        self.agent_positions = self._agent_positions[self._positions_in_use]
-        self.active_agents = self._agents[self._positions_in_use]
+    @property
+    def agent_positions(self):
+        """Return the positions of the agents in the space."""
+        return self._agent_positions[self._positions_in_use]
 
     def _remove_agent(self, agent: Agent) -> None:
         """Remove an agent from the space."""
@@ -140,7 +140,7 @@ class ContinuousSpace:
         self._index_to_agent.pop(index, None)
         self._positions_in_use[index] = False
         self._agents[index] = None
-        self._update_stuff()
+        self.active_agents = self._agents[self._positions_in_use]
 
     def calculate_difference_vector(self, point: np.ndarray, agents=None) -> np.ndarray:
         """Calculate the difference vector between the point and all agents."""
