### 2025-01-03 b7bdd82
make new continous spaces work with existing space drawing
