+++ mesa/experimental/continuous_space/__init__.py
@@ -1,3 +1,5 @@
+"""Continous space support."""
+
 from mesa.experimental.continuous_space.continuous_space import ContinuousSpace
 from mesa.experimental.continuous_space.continuous_space_agents import (
     ContinuousSpaceAgent,
