### 2025-01-03 eb5abd6
Update continuous_space.py
