### 2025-01-03 53ffa30
Update continuous_space.py
