+++ mesa/experimental/continuous_space/__init__.py
@@ -1,4 +1,4 @@
-"""Continous space support."""
+"""Continuous space support."""
 
 from mesa.experimental.continuous_space.continuous_space import ContinuousSpace
 from mesa.experimental.continuous_space.continuous_space_agents import (
