+++ mesa/experimental/continuous_space/continuous_space_agents.py
@@ -44,12 +44,12 @@ class ContinuousSpaceAgent(Agent):
 
     @property
     def pos(self):  # noqa: D102
-        # just here for compatability with solara_viz.
+        # just here for compatibility with solara_viz.
         return self.position
 
     @pos.setter
     def pos(self, value):
-        # just here for compatability solara_viz.
+        # just here for compatibility solara_viz.
         pass
 
     def __init__(self, space: ContinuousSpace, model):
