### 2025-01-04 6522c6f
Update release notes and version for 3.1.2
Update version and release notes for the next feature release, Mesa 3.1.2.
