### 2025-01-04 eee6f9b
creating sliders
sliders created for step interval to display graph after specified number of steps
