### 2025-01-04 e626ddf
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
