+++ mesa/mesa_logging.py
@@ -10,7 +10,6 @@ import logging
 from functools import wraps
 from logging import DEBUG, INFO
 
-
 __all__ = [
     "DEBUG",
     "DEFAULT_LEVEL",
