### 2025-01-04 a6e7c11
adding missing file in previous commit
forgot to commit changes ofsolara_viz.py in previous commit
