+++ mesa/visualization/solara_viz.py
@@ -216,8 +216,13 @@ def ModelController(
     @function_logger(__name__)
     def do_step():
         """Advance the model by one step."""
-        model.value.step()
+        
+        for _ in range(model_parameters.value['stepsize']):
+
+            model.value.step()
+
         running.value = model.value.running
+        
         force_update()
 
     @function_logger(__name__)
