### 2025-01-05 4136b1e
paper: Update space section - adding clarity.
