+++ mesa/examples/basic/boid_flockers/agents.py
@@ -62,8 +62,7 @@ class Boid(ContinuousSpaceAgent):
     def step(self):
         """Get the Boid's neighbors, compute the new vector, and move accordingly."""
         neighbors, distances = self.get_neighbors_in_radius(
-            radius=self.vision, include_distance=True
-        )
+            radius=self.vision)
         self.neighbors = neighbors
 
         # If no neighbors, maintain current direction
