+++ mesa/space.py
@@ -25,7 +25,6 @@ Classes
 # Remove this __future__ import once the oldest supported Python is 3.10
 from __future__ import annotations
 
-
 import collections
 import contextlib
 import inspect
@@ -1403,7 +1402,6 @@ class ContinuousSpace:
 
         self._invalidate_agent_cache()
         agent.pos = None
-
 
     def get_neighbors(
         self, pos: FloatCoordinate, radius: float, include_center: bool = True
