+++ tests/test_continuous_space.py
@@ -69,7 +69,7 @@ def test_continuous_agent():
     assert space.agent_positions.shape == (10, 2)
     for agent in space.agents:
         a = agent.position
-        b = space._agent_positions[space._get_index_for_agent(agent)]
+        b = space._agent_positions[space._agent_to_index[agent]]
         assert np.all(a == b)
 
     # add more agents, triggering a resizeing of the array
@@ -80,7 +80,7 @@ def test_continuous_agent():
     assert space.agent_positions.shape == (110, 2)
     for agent in space.agents:
         a = agent.position
-        b = space._agent_positions[space._get_index_for_agent(agent)]
+        b = space._agent_positions[space._agent_to_index[agent]]
         assert np.all(a == b)
 
     # remove all agents and check if the view is updated throughout correctly
