### 2025-01-05 102492a
switch active agents to list
