+++ tests/test_model.py
@@ -1,4 +1,5 @@
 """Tests for model.py."""
+
 import numpy as np
 
 from mesa.agent import Agent, AgentSet
@@ -39,7 +40,14 @@ def test_seed(seed=23):
     assert model._seed == seed
 
     assert Model(seed=42).random.random() == Model(seed=42).random.random()
-    assert np.all(Model(seed=42).rng.random(10,) == Model(seed=42).rng.random(10,))
+    assert np.all(
+        Model(seed=42).rng.random(
+            10,
+        )
+        == Model(seed=42).rng.random(
+            10,
+        )
+    )
 
 
 def test_reset_randomizer(newseed=42):
