### 2025-01-06 4683db9
Update space.py
