+++ mesa/space.py
@@ -1418,7 +1418,9 @@ class ContinuousSpace:
 
         Notes:
             If 1 or more agents are located on pos, include_center=True will remove these agents
-            from the results.
+            from the results. So, if you really want to get the neighbors of a given agent,
+            you should set include_center to true and filter the list of agents to remove
+            the given agent. 
 
         """
         if self._agent_points is None:
