### 2025-01-06 0e3460a
Update space.py
