+++ tests/test_continuous_space.py
@@ -204,16 +204,10 @@ def test_difference_vector():
     agent.position = [0.1, 0.1]
 
     vector = space.calculate_difference_vector([0.1, 0.9])
-    assert np.all(
-        vector
-        == [0, -0.8]
-    )
+    assert np.all(vector == [0, -0.8])
 
     vector = space.calculate_difference_vector([0.9, 0.1])
-    assert np.all(
-        vector
-        == [-0.8, 0]
-    )
+    assert np.all(vector == [-0.8, 0])
 
     # torus
     model = Model(seed=42)
@@ -231,6 +225,7 @@ def test_difference_vector():
 
     vector = space.calculate_difference_vector([0.9, 0.9])
     assert np.allclose(vector, [0.2, 0.2])
+
 
 # class TestSpaceToroidal(unittest.TestCase):
 #     """Testing a toroidal continuous space."""
