### 2025-01-06 c7df9be
Update test_continuous_space.py
