+++ tests/test_continuous_space.py
@@ -285,67 +285,52 @@ def test_continuous_space_get_k_nearest_agents():
 
 
 def test_continuous_space_get_agents_in_radius():
-    pass
+    # non torus
+    model = Model(seed=42)
+    dimensions = np.asarray([[0, 1], [0, 1]])
+    space = ContinuousSpace(dimensions, torus=False, random=model.random)
 
+    positions = [[0.1, 0.1],
+                 [0.1, 0.9],
+                 [0.9, 0.1],
+                 [0.9, 0.9],
+                 [0.5, 0.5],]
 
-# class TestSpaceToroidal(unittest.TestCase):
-#     """Testing a toroidal continuous space."""
-#
-#     def setUp(self):
-#         """Create a test space and populate with Mock Agents."""
-#         self.space = ContinuousSpace(70, 20, True, -30, -30)
-#         self.agents = []
-#         for i, pos in enumerate(TEST_AGENTS):
-#             a = MockAgent(i)
-#             self.agents.append(a)
-#             self.space.place_agent(a, pos)
-#
-#     def test_agent_positions(self):
-#         """Ensure that the agents are all placed properly."""
-#         for i, pos in enumerate(TEST_AGENTS):
-#             a = self.agents[i]
-#             assert a.pos == pos
-#
-#     def test_agent_matching(self):
-#         """Ensure that the agents are all placed and indexed properly."""
-#         for i, agent in self.space._index_to_agent.items():
-#             assert agent.pos == tuple(self.space._agent_points[i, :])
-#             assert i == self.space._agent_to_index[agent]
-#
-#     def test_distance_calculations(self):
-#         """Test toroidal distance calculations."""
-#         pos_1 = (-30, -30)
-#         pos_2 = (70, 20)
-#         assert self.space.get_distance(pos_1, pos_2) == 0
-#
-#         pos_3 = (-30, -20)
-#         assert self.space.get_distance(pos_1, pos_3) == 10
-#
-#         pos_4 = (20, -5)
-#         pos_5 = (20, -15)
-#         assert self.space.get_distance(pos_4, pos_5) == 10
-#
-#         pos_6 = (-30, -29)
-#         pos_7 = (21, -5)
-#         assert self.space.get_distance(pos_6, pos_7) == np.sqrt(49**2 + 24**2)
-#
-#     def test_heading(self):
-#         pos_1 = (-30, -30)
-#         pos_2 = (70, 20)
-#         self.assertEqual((0, 0), self.space.get_heading(pos_1, pos_2))
-#
-#         pos_1 = (65, -25)
-#         pos_2 = (-25, -25)
-#         self.assertEqual((10, 0), self.space.get_heading(pos_1, pos_2))
-#
-#     def test_neighborhood_retrieval(self):
-#         """Test neighborhood retrieval."""
-#         neighbors_1 = self.space.get_neighbors((-20, -20), 1)
-#         assert len(neighbors_1) == 2
-#
-#         neighbors_2 = self.space.get_neighbors((40, -10), 10)
-#         assert len(neighbors_2) == 0
-#
-#         neighbors_3 = self.space.get_neighbors((-30, -30), 10)
-#         assert len(neighbors_3) == 1
-#
+    for position in positions:
+        agent = ContinuousSpaceAgent(space, model)
+        agent.position = position
+
+    agents, distances = space.get_agents_in_radius([0.1, 0.1], radius=0.1)
+    assert len(agents) == 1
+    assert np.allclose(
+        distances,
+        [
+            0,
+        ],
+    )
+
+    agents, distances = space.get_agents_in_radius([0.5, 0.1], radius=0.4)
+    assert len(agents) == 3
[…]
