### 2025-01-06 436a589
Update mesa/space.py
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>
