+++ mesa/space.py
@@ -1417,7 +1417,7 @@ class ContinuousSpace:
                             agent in the results.
 
         Notes:
-            If 1 or more agents are located on pos, include_center=True will remove all these agents
+            If 1 or more agents are located on pos, include_center=False will remove all these agents
             from the results. So, if you really want to get the neighbors of a given agent,
             you should set include_center=True, and then filter the list of agents to remove
             the given agent (i.e., self when calling it from an agent).
