### 2025-01-06 3529b7d
Basic bugfix for bug ContinuousSpace.get_neighbors (#2599)
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>
