### 2025-01-06 6f0a46a
merge related fixes
