+++ mesa/examples/basic/boid_flockers/agents.py
@@ -61,7 +61,7 @@ class Boid(ContinuousSpaceAgent):
 
     def step(self):
         """Get the Boid's neighbors, compute the new vector, and move accordingly."""
-        neighbors = self.model.space.get_neighbors(self.pos, self.vision, True)
+        neighbors, distances = self.get_neighbors_in_radius(radius=self.vision)
         self.neighbors = [n for n in neighbors if n is not self]
 
         # If no neighbors, maintain current direction
