+++ mesa/experimental/continuous_space/continuous_space_agents.py
@@ -83,7 +83,7 @@ class ContinuousSpaceAgent(Agent):
         agents, dists = self.space.get_agents_in_radius(self.position, radius=radius)
 
         if dists.size == 0:
-            return dists, agents
+            return agents, dists
 
         logical = np.asarray([agent is not self for agent in agents])
         agents = list(compress(agents, logical))
