### 2025-01-08 7780b3a
typing and tests
