+++ mesa/examples/basic/boid_flockers/model.py
@@ -101,5 +101,5 @@ class BoidFlockers(Model):
 if __name__ == "__main__":
     model = BoidFlockers(population_size=200, width=100, height=100, vision=5, seed=42)
 
-    for _i in range(100):
+    for _i in range(20):
         model.step()
