+++ mesa/experimental/continuous_space/continuous_space.py
@@ -1,17 +1,16 @@
 """A Continuous Space class."""
 
 import warnings
-from collections.abc import Sequence
+from collections.abc import Iterable, Sequence
 from itertools import compress
 from random import Random
 
 import numpy as np
+from numpy.typing import ArrayLike
 from scipy.spatial.distance import cdist
 
 from mesa.agent import Agent, AgentSet
 
-from typing import Iterable, Sequence
-from numpy.typing import ArrayLike
 
 class ContinuousSpace:
     """Continuous space where each agent can have an arbitrary position."""
@@ -168,7 +167,9 @@ class ContinuousSpace:
 
         return delta
 
-    def calculate_distances(self, point:ArrayLike, agents:Iterable[Agent]|None=None) -> tuple[np.ndarray, list]:
+    def calculate_distances(
+        self, point: ArrayLike, agents: Iterable[Agent] | None = None
+    ) -> tuple[np.ndarray, list]:
         """Calculate the distance between the point and all agents."""
         point = np.asanyarray(point)
 
@@ -192,7 +193,9 @@ class ContinuousSpace:
             dists = cdist(point[np.newaxis, :], positions)[0, :]
         return dists, agents
 
-    def get_agents_in_radius(self, point:ArrayLike, radius:float|int=1) -> tuple[list, np.ndarray]:
+    def get_agents_in_radius(
+        self, point: ArrayLike, radius: float | int = 1
+    ) -> tuple[list, np.ndarray]:
         """Return the agents and their distances within a radius for the point."""
         distances, agents = self.calculate_distances(point)
         logical = distances <= radius
@@ -202,7 +205,9 @@ class ContinuousSpace:
             distances[logical],
         )
 
-    def get_k_nearest_agents(self, point:ArrayLike, k:int=1) -> tuple[list, np.ndarray]:
+    def get_k_nearest_agents(
+        self, point: ArrayLike, k: int = 1
+    ) -> tuple[list, np.ndarray]:
         """Return the k nearest agents and their distances to the point.
 
         Notes:
