+++ tests/test_continuous_space.py
@@ -103,15 +103,14 @@ def test_continuous_agent():
     dimensions = np.asarray([[0, 1], [0, 1]])
     space = ContinuousSpace(dimensions, torus=True, random=model.random)
     agent = ContinuousSpaceAgent(space, model)
-    agent.position =[1.1, 1.1]
+    agent.position = [1.1, 1.1]
     assert np.allclose(agent.position, [0.1, 0.1])
 
     dimensions = np.asarray([[0, 1], [0, 1]])
     space = ContinuousSpace(dimensions, torus=False, random=model.random)
     agent = ContinuousSpaceAgent(space, model)
     with pytest.raises(ValueError):
-        agent.position =[1.1, 1.1]
-
+        agent.position = [1.1, 1.1]
 
 
 def test_continous_space_calculate_distances():
