### 2025-01-08 c4de8de
Update mesa/examples/basic/boid_flockers/model.py
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>
