### 2025-01-08 23159bf
Update mesa/examples/basic/boid_flockers/model.py
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>
