### 2025-01-08 0706667
expanded docstring and comments
