### 2025-01-08 4187786
Update continuous_space.py
