+++ mesa/experimental/continuous_space/continuous_space.py
@@ -74,7 +74,7 @@ class ContinuousSpace:
             random = Random()
         self.random = random
 
-        self.dimensions: np.array = np.asanarray(dimensions)
+        self.dimensions: np.array = np.asanyarray(dimensions)
         self.ndims: int = self.dimensions.shape[0]
         self.size: np.array = self.dimensions[:, 1] - self.dimensions[:, 0]
         self.center: np.array = np.sum(self.dimensions, axis=1) / 2
