### 2025-01-09 #7620 @gkthiruvathukal
Hi @danielskatz! Happy new year! I'm happy to help.
