### 2025-01-09 #7620 @gkthiruvathukal
Hi @jamesdamillington and @martibosch, as you have both been involved in a previous review for a Python package based on agent-based modeling, would you be willing to contribute a review for this JOSS submission?
