### 2025-01-09 51022e3
Added permanent step interval slider to SolaraViz for precise adjustments
