+++ mesa/examples/basic/boid_flockers/app.py
@@ -17,13 +17,6 @@ model_params = {
         "value": 42,
         "label": "Random Seed",
     },
-    "stepsize": Slider(
-        label="Steps Interval",
-        value=1,
-        min=1,
-        max=20,
-        step=1,
-    ),
     "population": Slider(
         label="Number of boids",
         value=100,
