+++ mesa/examples/advanced/sugarscape_g1mt/app.py
@@ -48,13 +48,6 @@ model_params = {
         "value": 42,
         "label": "Random Seed",
     },
-    "stepsize": Slider(
-        label="Steps Interval",
-        value=1,
-        min=1,
-        max=20,
-        step=1,
-    ),
     "width": 50,
     "height": 50,
     # Population parameters
