+++ mesa/examples/basic/boid_flockers/model.py
@@ -27,7 +27,6 @@ class BoidFlockers(Model):
         separate=0.015,
         match=0.05,
         seed=None,
-        stepsize=1,
     ):
         """Create a new Boids Flocking model.
 
