+++ mesa/examples/advanced/epstein_civil_violence/app.py
@@ -57,7 +57,6 @@ model_params = {
     "cop_vision": Slider("Cop Vision", 7, 1, 10, 1),
     "legitimacy": Slider("Government Legitimacy", 0.82, 0.0, 1, 0.01),
     "max_jail_term": Slider("Max Jail Term", 30, 0, 50, 1),
-    "stepsize": 1,
 }
 
 space_component = make_space_component(
