+++ mesa/examples/advanced/epstein_civil_violence/model.py
@@ -48,7 +48,6 @@ class EpsteinCivilViolence(mesa.Model):
         movement=True,
         max_iters=1000,
         seed=None,
-        stepsize=1,
     ):
         super().__init__(seed=seed)
         self.movement = movement
