### 2025-01-09 8d4a985
Revert unintended changes
Revert unintended changes to mesa\visualization\components\__init__.py
