+++ mesa/experimental/continuous_space/continuous_space_agents.py
@@ -67,7 +67,7 @@ class ContinuousSpaceAgent(Agent):
         # self.position[:] = np.nan
 
     def remove(self) -> None:
-        """Remove and delete the agent from the model and continuous space"""
+        """Remove and delete the agent from the model and continuous space."""
         super().remove()
         self.space._remove_agent(self)
         self._mesa_index = None
