### 2025-01-10 897d62e
Update continuous_space.py
