### 2025-01-10 70375d1
Reimplementation of Continuous Space (#2584)
