### 2025-01-10 29d0f3b
Reimplementation of Continuous Space (#2584)
