+++ tests/test_continuous_space.py
@@ -1,0 +1,432 @@
+"""Tests for continuous space."""
+
+import numpy as np
+import pytest
+
+from mesa import Model
+from mesa.experimental.continuous_space import ContinuousSpace, ContinuousSpaceAgent
+
+
+def test_continuous_space():
+    """Test ContinuousSpace class."""
+    model = Model(seed=42)
+
+    dimensions = np.asarray([[0, 1], [-1, 0]])
+    space = ContinuousSpace(dimensions, torus=False, random=model.random)
+
+    # check some default fields
+    assert space.ndims == 2
+    assert np.all(space.size == [1, 1])
+    assert np.all(space.center == [0.5, -0.5])
+    assert len(space.agents) == 0
+
+    # check in_bounds
+    assert space.in_bounds([0.5, -0.5])
+    assert not space.in_bounds([-0.5, -0.5])
+    assert not space.in_bounds([1.5, -0.5])
+    assert not space.in_bounds([0.5, 0.5])
+    assert not space.in_bounds([0.5, -1.5])
+
+    # check torus correction
+    space = ContinuousSpace(dimensions, torus=True, random=model.random)
+    assert np.all(space.torus_correct([-0.5, 0.5]) == [0.5, -0.5])
+    assert np.all(space.torus_correct([0.5, -0.5]) == [0.5, -0.5])
+    assert np.all(space.torus_correct([0.5, -0.5]) == [0.5, -0.5])
+    assert np.all(space.torus_correct([1.5, -0.5]) == [0.5, -0.5])
+    assert np.all(space.torus_correct([0.5, -1.5]) == [0.5, -0.5])
+
+    # check 3d
+    dimensions = np.asarray([[0, 2], [-2, 0], [-2, 2]])
+    space = ContinuousSpace(dimensions, torus=False, random=model.random)
+
+    # check some default fields
+    assert space.ndims == 3
+    assert np.all(space.size == [2, 2, 4])
+    assert np.all(space.center == [1, -1, 0])
+
+    # check in_bounds
+    assert space.in_bounds([1, -1, 0])
+    assert not space.in_bounds([-0.5, -1, 0])
+    assert not space.in_bounds([2.5, -1, 0])
+    assert not space.in_bounds([1, 0.5, 0])
+    assert not space.in_bounds([1, -2.5, 0])
+    assert not space.in_bounds([1, -1, -3])
+    assert not space.in_bounds([1, -1, 3])
+    assert not space.in_bounds([-0.5, -1, 3])
+    assert not space.in_bounds([1, 0.5, 3])
+
+
+def test_continuous_agent():
+    """Test ContinuousSpaceAgent class."""
+    model = Model(seed=42)
+
+    dimensions = np.asarray([[0, 1], [0, 1]])
+    space = ContinuousSpace(dimensions, torus=False, random=model.random)
+
+    for _ in range(10):
+        agent = ContinuousSpaceAgent(space, model)
+        position = [agent.random.random(), agent.random.random()]
+        agent.position = position
+        agent.coordinate = position
+
+    assert space.agent_positions.shape == (10, 2)
+    for agent in space.agents:
+        a = agent.position
+        b = space._agent_positions[space._agent_to_index[agent]]
+        assert np.all(a == b)
+        assert np.all(agent.position == agent.coordinate)
+
+    # add more agents, triggering a resizeing of the array
+    for _ in range(100):
+        agent = ContinuousSpaceAgent(space, model)
+        position = [agent.random.random(), agent.random.random()]
+        agent.position = position
+        agent.coordinate = position
+
+    assert space.agent_positions.shape == (110, 2)
+    for agent in space.agents:
+        a = agent.position
[…]
