### 2025-01-10 b4c71e4
Rename step_interval to render_interval for clarity and consistency
