+++ mesa/visualization/solara_viz.py
@@ -52,7 +52,7 @@ def SolaraViz(
     | Literal["default"] = "default",
     *,
     play_interval: int = 100,
-    step_interval: int = 1,
+    render_interval: int = 1,
     simulator: Simulator | None = None,
     model_params=None,
     name: str | None = None,
@@ -73,7 +73,7 @@ def SolaraViz(
             Defaults to "default", which uses the default Altair space visualization.
         play_interval (int, optional): Interval for playing the model steps in milliseconds.
             This controls the speed of the model's automatic stepping. Defaults to 100 ms.
-        step_interval (int, optional): Controls how often plots are updated during a simulation,
+        render_interval (int, optional): Controls how often plots are updated during a simulation,
             allowing users to skip intermediate steps and update graphs less frequently.
         simulator: A simulator that controls the model (optional)
         model_params (dict, optional): Parameters for (re-)instantiating a model.
@@ -93,7 +93,7 @@ def SolaraViz(
           model instance is provided, it will be converted to a reactive model using `solara.use_reactive`.
         - The `play_interval` argument controls the speed of the model's automatic stepping. A lower
           value results in faster stepping, while a higher value results in slower stepping.
-        - The `step_interval` argument determines how often plots are updated during simulation. Higher values
+        - The `render_interval` argument determines how often plots are updated during simulation. Higher values
           reduce update frequency,resulting in faster execution.
     """
     if components == "default":
@@ -108,7 +108,7 @@ def SolaraViz(
     # set up reactive model_parameters shared by ModelCreator and ModelController
     reactive_model_parameters = solara.use_reactive({})
     reactive_play_interval = solara.use_reactive(play_interval)
-    reactive_step_interval = solara.use_reactive(step_interval)
+    reactive_render_interval = solara.use_reactive(render_interval)
     with solara.AppBar():
         solara.AppBarTitle(name if name else model.value.__class__.__name__)
 
@@ -123,9 +123,9 @@ def SolaraViz(
                 step=10,
             )
             solara.SliderInt(
-                label="Step Interval",
-                value=reactive_step_interval,
-                on_value=lambda v: reactive_step_interval.set(v),
+                label="Render Interval",
+                value=reactive_render_interval,
+                on_value=lambda v: reactive_render_interval.set(v),
                 min=1,
                 max=100,
                 step=2,
@@ -135,7 +135,7 @@ def SolaraViz(
                     model,
                     model_parameters=reactive_model_parameters,
                     play_interval=reactive_play_interval,
-                    step_interval=reactive_step_interval,
+                    render_interval=reactive_render_interval,
                 )
             else:
                 SimulatorController(
@@ -143,7 +143,7 @@ def SolaraViz(
                     simulator,
                     model_parameters=reactive_model_parameters,
                     play_interval=reactive_play_interval,
-                    step_interval=reactive_step_interval,
+                    render_interval=reactive_render_interval,
                 )
         with solara.Card("Model Parameters"):
             ModelCreator(
@@ -204,7 +204,7 @@ def ModelController(
     *,
     model_parameters: dict | solara.Reactive[dict] = None,
     play_interval: int | solara.Reactive[int] = 100,
-    step_interval: int | solara.Reactive[int] = 1,
+    render_interval: int | solara.Reactive[int] = 1,
 ):
     """Create controls for model execution (step, play, pause, reset).
 
@@ -212,7 +212,7 @@ def ModelController(
         model: Reactive model instance
         model_parameters: Reactive parameters for (re-)instantiating a model.
         play_interval: Interval for playing the model steps in milliseconds.
-        step_interval: Controls how often the plots are updated during simulation steps.Higher value reduce update frequency.
[…]
