### 2025-01-10 48aef48
reimplementation of draw_voroinoi
