### 2025-01-10 6bce2c8
Remove changes to examples/advanced/pd_grid/model.py
