### 2025-01-10 5b90f20
Remove bullet
