### 2025-01-10 baf5c87
reimplementation of draw_voroinoi (#2608)
* reimplementation of draw_voroinoi

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
