### 2025-01-10 f716287
fixed hex-space draw to avoid overlaps
