### 2025-01-10 f17e99d
Add Polygon and PatchCollection imports for visualization
