+++ mesa/visualization/mpl_space_drawing.py
@@ -12,7 +12,6 @@ import warnings
 from collections.abc import Callable
 from typing import Any
 
-from matplotlib.patches import Polygon
 import networkx as nx
 import numpy as np
 from matplotlib import pyplot as plt
@@ -20,6 +19,7 @@ from matplotlib.axes import Axes
 from matplotlib.cm import ScalarMappable
 from matplotlib.collections import LineCollection, PatchCollection
 from matplotlib.colors import LinearSegmentedColormap, Normalize, to_rgba
+from matplotlib.patches import Polygon
 
 import mesa
 from mesa.experimental.cell_space import (
