+++ mesa/examples/advanced/sugarscape_g1mt/model.py
@@ -3,9 +3,9 @@ from pathlib import Path
 import numpy as np
 
 import mesa
-from mesa.examples.advanced.sugarscape_g1mt.agents import Trader
 from mesa.discrete_space import OrthogonalVonNeumannGrid
 from mesa.discrete_space.property_layer import PropertyLayer
+from mesa.examples.advanced.sugarscape_g1mt.agents import Trader
 
 
 # Helper Functions
