+++ mesa/examples/advanced/wolf_sheep/model.py
@@ -13,8 +13,8 @@ import math
 
 from mesa import Model
 from mesa.datacollection import DataCollector
-from mesa.examples.advanced.wolf_sheep.agents import GrassPatch, Sheep, Wolf
 from mesa.discrete_space import OrthogonalVonNeumannGrid
+from mesa.examples.advanced.wolf_sheep.agents import GrassPatch, Sheep, Wolf
 from mesa.experimental.devs import ABMSimulator
 
 
