+++ mesa/experimental/cell_space/__init__.py
@@ -51,8 +51,6 @@ __all__ = [
     "VoronoiGrid",
 ]
 
-warnings.warn(
-    "you are importing from mesa.discrete_space,"
-    "all cell spaces have been moved to mesa.discrete_space",
-    DeprecationWarning,
-)
+warnings.warn("you are importing from mesa.discrete_space,"
+              "all cell spaces have been moved to mesa.discrete_space", DeprecationWarning,
+              stacklevel=2,)
