+++ docs/apis/experimental.md
@@ -1,47 +1,6 @@
 # Experimental
 This namespace contains experimental features. These are under development, and their API is not necessarily stable.
 
-## Cell Space
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.__init__
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.cell
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.cell_agent
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.cell_collection
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.discrete_space
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.grid
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.network
-   :members:
-```
-
-```{eval-rst}
-.. automodule:: experimental.cell_space.voronoi
-   :members:
-```
 
 ## Devs
 
