+++ docs/apis/discrete_space.md
@@ -1,0 +1,41 @@
+## Discrete Space
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.__init__
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.cell
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.cell_agent
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.cell_collection
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.discrete_space
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.grid
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.network
+   :members:
+```
+
+```{eval-rst}
+.. automodule:: experimental.cell_space.voronoi
+   :members:
+```
