### 2025-01-11 eea1c6a
Update discrete_space.md
