### 2025-01-11 db0225b
Add docs for continuous space (#2613)
