### 2025-01-12 #7620 @EwoutH
@gkthiruvathukal thanks for picking up as editor! If you need anything from our side let me know.

Since I don’t think anyone has properly tagged @jofmi before, I’m curious if you (and @martibosch) are willing to review our Mesa 3 paper!
