+++ mesa/experimental/cell_space/__init__.py
@@ -51,6 +51,7 @@ __all__ = [
     "VoronoiGrid",
 ]
 
+
 warnings.warn(
     "you are importing from mesa.experimental.cell_space, "
     "all cell spaces have been moved to mesa.discrete_space",
