### 2025-01-12 cad6131
Move grass regrowth time to PropertyLayer
