### 2025-01-12 7589b6c
Rename wolf sheep constructor parameter name
