### 2025-01-12 9f24d13
paper: Update to visualization section.
