+++ paper/paper.md
@@ -102,7 +102,7 @@ Mesa 3 provides both discrete (cell-based) and continuous space implementations.
 Example grid creation:
 
 ```python
-   grid = OrthogonalVonNeumannGrid((width, height), torus=False, random=model.random)
+    grid = OrthogonalVonNeumannGrid((width, height), torus=False, random=model.random)
 ```
 
 In Mesa 3, specialized agent classes for spatial interactions in discrete spaces were added:
@@ -114,15 +114,15 @@ In Mesa 3, specialized agent classes for spatial interactions in discrete spaces
 All discrete spaces support PropertyLayers - efficient numpy-based arrays for storing cell-level properties. This newly added feature allows for agents to interact with spatial properties of the cell more easily:
 
 ```python
-   grid.create_property_layer("elevation", default_value=10)
-   high_ground = grid.elevation.select_cells(lambda x: x > 50)
+    grid.create_property_layer("elevation", default_value=10)
+    high_ground = grid.elevation.select_cells(lambda x: x > 50)
 ```
 
 For models where agents need to move continuously through space rather than between discrete locations, `ContinuousSpace` allows agents to occupy any coordinate within defined boundaries:
 
 ```python
-   space = ContinuousSpace(x_max, y_max, torus=True)
-   space.move_agent(agent, (new_x, new_y))
+    space = ContinuousSpace(x_max, y_max, torus=True)
+    space.move_agent(agent, (new_x, new_y))
 ```
 The space module is stable but under active development. The new cell-based spaces in Mesa 3 are currently being tested and considered feature-complete. The code snippets reflected in this paper are the future expected state of Mesa. Features not yet merged into core are imported from experimental:
 
@@ -134,61 +134,60 @@ from mesa.experimental.cell_space ...
 Typically, ABMs represent time incrementally and call the units ticks. For each tick, the step method of the model is called, and the agents are activated to take their designated actions. The most frequently implemented approach is shown below, which runs a model for 100 ticks.
 
 ```python
-   model = Model(seed=42)
+    model = Model(seed=42)
 
-
-   for _ in range(100):
-       model.step()
+    for _ in range(100):
+        model.step()
 ```
 
-Before Mesa 3, all agents were activated within the step method of the model. However, the newly added `AgentSet` class provides a more flexible way to activate agents. These changes include the depreciation of the Scheduler API and replacing previously available fixed patterns. 
-
+Before Mesa 3, all agents were activated within the step method of the model. However, the newly added `AgentSet` class provides a more flexible way to activate agents. These changes include the depreciation of the Scheduler API and replacing previously available fixed patterns.
 
 ```python
-   model.agents.do("step")  # Sequential activation
+    model.agents.do("step")  # Sequential activation
 
-   model.agents.shuffle_do("step")  # Random activation
+    model.agents.shuffle_do("step")  # Random activation
 
-   # Multi-stage activation:
-   for stage in ["move", "eat", "reproduce"]:
-       model.agents.do(stage)
+    # Multi-stage activation:
+    for stage in ["move", "eat", "reproduce"]:
+        model.agents.do(stage)
 
-   # Activation by agent subclass:
-   for klass in model.agent_types:
-       model.agents_by_type[klass].do("step")
+    # Activation by agent subclass:
+    for klass in model.agent_types:
+        model.agents_by_type[klass].do("step")
 ```
 
[…]
