### 2025-01-14 be42172
Update paper/paper.md
Co-authored-by: Ewout ter Hoeven <E.M.terHoeven@student.tudelft.nl>
