+++ paper/paper.bib
@@ -105,15 +105,16 @@
   doi = {10.1016/j.enbuild.2020.109763}
 }
 
-@article{anatolitis2017putting,
-  title={Putting renewable energy auctions into action--An agent-based model of onshore wind power auctions in Germany},
-  author={Anatolitis, Vasilios and Welisch, Marijke},
-  journal={Energy Policy},
-  volume={110},
-  pages={394--402},
-  year={2017},
-  publisher={Elsevier},
-  doi = {10.1016/j.enpol.2017.08.024}
+@article{pham2021interventions,,
+	title = {Interventions to control nosocomial transmission of {SARS}-{CoV}-2: a modelling study},
+	author = {Pham, Thi Mui and Tahir, Hannan and van de Wijgert, Janneke H. H. M. and Van der Roest, Bastiaan R. and Ellerbroek, Pauline and Bonten, Marc J. M. and Bootsma, Martin C. J. and Kretzschmar, Mirjam E.},
+	journal = {BMC Medicine},
+	volume = {19},
+	pages = {211},
+	year = {2021},
+  month = aug,
+  publisher={Springer},
+	doi = {10.1186/s12916-021-02060-y}
 }
 
 # Used Mesa (cited 2020 paper)
@@ -139,14 +140,12 @@
   doi = {10.1016/j.elerap.2022.101195}
 }
 
-@article{taberna2023uncertainty,
-  title={Uncertainty in boundedly rational household adaptation to environmental shocks},
-  author={Taberna, Alessandro and Filatova, Tatiana and Hadjimichael, Antonia and Noll, Brayton},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={120},
-  number={44},
-  pages={e2215675120},
-  year={2023},
-  publisher={National Acad Sciences},
-  doi = {10.1073/pnas.2215675120}
+@article{souza2023edgesimpy,
+  title = {EdgeSimPy: Python-based modeling and simulation of edge computing resource management policies},
+  author = {Paulo S. Souza and Tiago Ferreto and Rodrigo N. Calheiros},
+  journal = {Future Generation Computer Systems},
+  volume = {148},
+  pages = {446-459},
+  year = {2023},
+  doi = {https://doi.org/10.1016/j.future.2023.06.013}
 }
