+++ paper/paper.md
@@ -232,10 +232,10 @@ Mesa has been applied across diverse domains, including:
 - Sustainability and food security modeling, including supply chain optimization and resource allocation [@namany2020sustainable]
 - Infrastructure resilience and post-disaster recovery planning [@sun2020post]
 - Building stock modeling for energy demand and emissions analysis [@nageli2020towards]
-- Market modeling, including renewable energy auctions and consumer behavior [@anatolitis2017putting]
 - Transportation optimization, such as combined truck-drone delivery routing [@leon2022multi]
 - Recommender systems analysis examining consumer-business value tradeoffs over time [@ghanem2022balancing]
-- Climate adaptation modeling examining household-level behavioral responses to environmental shocks [@taberna2023uncertainty]
+- SEIR modeling of Sars-CoV-2 (Covid-19) [@pham2021interventions]
+- Management of edge computing resources [@souza2023edgesimpy]
 
 These applications showcase Mesa's versatility in modeling complex systems with autonomous interacting agents, whether representing individual consumers, infrastructure components, buildings, or vehicles.
 
