### 2025-01-15 e6a1b34
paper: Edit research citations mentioned.
