+++ paper/paper.bib
@@ -72,17 +72,6 @@
 }
 
 # Used Mesa (cited 2015 paper)
-@article{namany2020sustainable,
-  title={Sustainable food security decision-making: an agent-based modelling approach.},
-  author={Namany, Sarah and Govindan, Rajesh and Alfagih, Luluwah and McKay, Gordon and Al-Ansari, Tareq},
-  journal={Journal of Cleaner Production},
-  volume={255},
-  pages={120296},
-  year={2020},
-  publisher={Elsevier},
-  doi = {10.1016/j.jclepro.2020.120296}
-}
-
 @article{sun2020post,
   title={A post-disaster resource allocation framework for improving resilience of interdependent infrastructure networks},
   author={Sun, Jingran and Zhang, Zhanmin},
@@ -94,17 +83,6 @@
   doi = {10.1016/j.trd.2020.102455}
 }
 
-@article{nageli2020towards,
-  title={Towards agent-based building stock modeling: Bottom-up modeling of long-term stock dynamics affecting the energy and climate impact of building stocks},
-  author={N{\"a}geli, Claudio and Jakob, Martin and Catenazzi, Giacomo and Ostermeyer, York},
-  journal={Energy and Buildings},
-  volume={211},
-  pages={109763},
-  year={2020},
-  publisher={Elsevier},
-  doi = {10.1016/j.enbuild.2020.109763}
-}
-
 @article{anatolitis2017putting,
   title={Putting renewable energy auctions into action--An agent-based model of onshore wind power auctions in Germany},
   author={Anatolitis, Vasilios and Welisch, Marijke},
@@ -114,6 +92,18 @@
   year={2017},
   publisher={Elsevier},
   doi = {10.1016/j.enpol.2017.08.024}
+}
+
+@article{pham2021interventions,
+	title = {Interventions to control nosocomial transmission of {SARS}-{CoV}-2: a modelling study},
+	author = {Pham, Thi Mui and Tahir, Hannan and van de Wijgert, Janneke H. H. M. and Van der Roest, Bastiaan R. and Ellerbroek, Pauline and Bonten, Marc J. M. and Bootsma, Martin C. J. and Kretzschmar, Mirjam E.},
+	journal = {BMC Medicine},
+	volume = {19},
+	pages = {211},
+	year = {2021},
+  month = aug,
+  publisher={Springer},
+	doi = {10.1186/s12916-021-02060-y}
 }
 
 # Used Mesa (cited 2020 paper)
@@ -128,6 +118,17 @@
   doi = {10.1016/j.eswa.2022.116604}
 }
 
+@article{taberna2023uncertainty,
+  title={Putting renewable energy auctions into action--An agent-based model of onshore wind power auctions in Germany},
+  author={Anatolitis, Vasilios and Welisch, Marijke},
+  journal={Energy Policy},
+  volume={110},
+  pages={394--402},
+  year={2017},
+  publisher={Elsevier},
+  doi = {10.1016/j.enpol.2017.08.024}
+}
+
 @article{ghanem2022balancing,
   title={Balancing consumer and business value of recommender systems: A simulation-based analysis},
   author={Ghanem, Nada and Leitner, Stephan and Jannach, Dietmar},
@@ -139,14 +140,12 @@
   doi = {10.1016/j.elerap.2022.101195}
 }
 
-@article{taberna2023uncertainty,
-  title={Uncertainty in boundedly rational household adaptation to environmental shocks},
-  author={Taberna, Alessandro and Filatova, Tatiana and Hadjimichael, Antonia and Noll, Brayton},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={120},
-  number={44},
-  pages={e2215675120},
-  year={2023},
[…]
