+++ .github/workflows/build_lint.yml
@@ -84,4 +84,4 @@ jobs:
     - name: Test examples
       run: |
         cd mesa-examples
-        pytest -rA -Werror -Wdefault::FutureWarning -Wi:DeprecationWarning test_examples.py
+        pytest -rA -Werror -Wdefault::FutureWarning -Wi::DeprecationWarning test_examples.py
