+++ docs/index.md
@@ -54,6 +54,7 @@ For help getting started with Mesa, check out these resources:
 - [Getting started] - Learn about Mesa's core concepts and components
 - [Migration Guide] - Upgrade to Mesa 3.0
 - [Mesa Examples] - Browse user-contributed models and implementations
+- [Mesa Extensions] - Overview of mesa's Extentions
 - [GitHub Discussions] - Ask questions and discuss Mesa
 - [Matrix Chat Room] - Real-time chat with the Mesa community
 
@@ -96,4 +97,4 @@ API Documentation <apis/api_main>
 [mesa visualization tutorial]: tutorials/visualization_tutorial
 [migration guide]: migration_guide
 [Getting started]: getting_started
-
+[Mesa Extensions]: mesa_extension.md
