### 2025-01-19 4722447
Optimize UI responsiveness by Offloading model execution to separate thread
