### 2025-01-20 0a34e2f
added empty loc check
