### 2025-01-20 #7668 @EwoutH
Awesome! If you very quickly want to run some examples, we also have an online demo:
- https://py.cafe/app/EwoutH/mesa-solara-basic-examples
