+++ docs/apis/visualization.md
@@ -11,6 +11,13 @@ For a detailed tutorial, please refer to our [Visualization Tutorial](../tutoria
    :show-inheritance:
 ```
 
+```{eval-rst}
+.. automodule:: mesa.visualization.components.__init__
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
 ## User Parameters
 
 ```{eval-rst}
@@ -37,12 +44,7 @@ For a detailed tutorial, please refer to our [Visualization Tutorial](../tutoria
    :show-inheritance:
 ```
 
-```{eval-rst}
-.. automodule:: mesa.visualization.components.__init__
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
+
 ## Altair-based visualizations
 
 ```{eval-rst}
