+++ mesa/visualization/__init__.py
@@ -13,10 +13,14 @@ from mesa.visualization.mpl_space_drawing import (
 from .components import make_plot_component, make_space_component
 from .components.altair_components import make_space_altair
 from .solara_viz import JupyterViz, SolaraViz
+from .types import HexGrid, Network, OrthogonalGrid
 from .user_param import Slider
 
 __all__ = [
+    "HexGrid",
     "JupyterViz",
+    "Network",
+    "OrthogonalGrid",
     "Slider",
     "SolaraViz",
     "draw_space",
