### 2025-01-21 33cbd9d
Remove version specific docs
Update version references in docs
to be more generic, avoiding the need to update them with each release.

Also, SolaraViz has been stable and well-tested through the Mesa 3.0.x series. Remove
the experimental warnings and version-specific notes from documentation, as it
is now a stable component of Mesa 3.1.
