+++ mesa/visualization/__init__.py
@@ -1,10 +1,4 @@
-"""Solara based visualization for Mesa models.
-
-.. note::
-    SolaraViz is experimental and still in active development for Mesa 3.0. While we attempt to minimize them, there might be API breaking changes between Mesa 3.0 and 3.1.
-
-    There won't be breaking changes between Mesa 3.0.x patch releases.
-"""
+"""Solara based visualization for Mesa models."""
 
 from mesa.visualization.mpl_space_drawing import (
     draw_space,
