### 2025-01-22 ac55736
adding thread without implementing render_interval
