### 2025-01-23 #7668 @jackiekazil
There was a minor change to the research mentioned in the paper.
No other changes to the paper are expected. It seems to me that there is some sort of publishing automation that JOSS has that needs to be triggered. Can you trigger it?
