### 2025-01-23 48d1bed
Update mpl_space_drawing.py
