### 2025-01-23 4a17ece
Update mpl_space_drawing.py
