### 2025-01-23 1ea215d
Update mpl_space_drawing.py
