### 2025-01-23 9a165eb
Update mpl_space_drawing.py
