### 2025-01-23 6eb5ea7
added conditional checks for empty grid placement
