### 2025-01-25 2561df8
render _interval functioning
execution is slower,thread not stopping gracefully
