### 2025-01-25 00c3c93
Update solara_viz.py
pending-
eipstein not rendering prply(check other models)
stopping thread(use_thread)
graph inaccuracy
