### 2025-01-25 a68b4fd
Update solara_viz.py
