### 2025-01-25 a08bdcb
removing asynchornous task to avoid cancelledError exception
