### 2025-01-25 fd17584
bidirectional synchronization
implementing bi-directional threads is ensuring synchronization between model execution and visualization, though this is resulting in reduced performance.
