### 2025-01-26 566b72f
renamed make_space_altair
