### 2025-01-26 e1f947c
added make_altair_plot_component support
