+++ mesa/visualization/components/altair_components.py
@@ -1,9 +1,10 @@
 """Altair based solara components for visualization mesa spaces."""
 
 import contextlib
+from typing import Callable
 import warnings
-
 import solara
+
 
 with contextlib.suppress(ImportError):
     import altair as alt
@@ -54,6 +55,105 @@ def make_altair_space(
 
     return MakeSpaceAltair
 
+
+
+def make_altair_plot_component(
+    measure: str | dict[str, str] | list[str] | tuple[str],
+    post_process: Callable | None = None,
+    width: int = 500,
+    height: int = 300,
+):
+    """Create an Altair plotting component for specified measures.
+
+    Args:
+        measure: Measure(s) to plot. Can be:
+            - str: Single measure name
+            - dict: Mapping of measure names to colors
+            - list/tuple: Multiple measure names
+        post_process: Optional callable for chart post-processing
+        width: Chart width in pixels
+        height: Chart height in pixels
+
+    Returns:
+        function: A function that creates a PlotAltair component
+    """
+    def MakePlotAltair(model):
+        return PlotAltair(
+            model,
+            measure,
+            post_process=post_process,
+            width=width,
+            height=height
+        )
+    return MakePlotAltair
+
+
+@solara.component
+def PlotAltair(
+    model,
+    measure: str | dict[str, str] | list[str] | tuple[str],
+    post_process: Callable[[alt.Chart], alt.Chart] | None = None,
+    width: int = 500,
+    height: int = 300,
+) -> solara.FigureAltair:
+    """Create an Altair plot for model."""
+    
+    update_counter.get()
+    df = model.datacollector.get_model_vars_dataframe().reset_index()
+
+    if isinstance(measure, str):
+        # Single measure - no transformation needed
+        chart = alt.Chart(df).encode(
+            x='Step:Q',
+            y=alt.Y(f'{measure}:Q', title=measure),
+            tooltip=[alt.Tooltip('Step:Q'), alt.Tooltip(f'{measure}:Q')]
+        ).mark_line()
+
+    elif isinstance(measure, (list, tuple)):
+        # Multiple measures - melt dataframe
+        value_vars = list(measure)
+        melted_df = df.melt('Step', value_vars=value_vars, 
+                           var_name='Measure', value_name='Value')
+        
+        chart = alt.Chart(melted_df).encode(
+            x='Step:Q',
+            y=alt.Y('Value:Q'),
+            color='Measure:N',
+            tooltip=['Step:Q', 'Value:Q', 'Measure:N']
+        ).mark_line()
+
+    elif isinstance(measure, dict):
+        # Dictionary with colors - melt dataframe
+        value_vars = list(measure.keys())
+        melted_df = df.melt('Step', value_vars=value_vars,
+        var_name='Measure', value_name='Value')
+        
+        # Create color scale from measure dict
+        domain = list(measure.keys())
+        range_ = list(measure.values())
+        
+        chart = alt.Chart(melted_df).encode(
+            x='Step:Q',
+            y=alt.Y('Value:Q'),
+            color=alt.Color('Measure:N', scale=alt.Scale(domain=domain, range=range_)),
+            tooltip=['Step:Q', 'Value:Q', 'Measure:N']
+        ).mark_line()
+    
+    else:
+        raise ValueError("Unsupported measure type")
+
+    # Configure chart properties
+    chart = chart.properties(
+        width=width,
+        height=height
+    ).configure_axis(
+        grid=True
+    )
+
+    if post_process is not None:
+        chart = post_process(chart)
+        
+    return solara.FigureAltair(chart)
 
 @solara.component
 def SpaceAltair(model, agent_portrayal, dependencies: list[any] | None = None):
