### 2025-01-26 9e3a625
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
