+++ mesa/visualization/__init__.py
@@ -11,7 +11,7 @@ from mesa.visualization.mpl_space_drawing import (
 )
 
 from .components import make_plot_component, make_space_component
-from .components.altair_components import make_space_altair,make_altair_plot_component
+from .components.altair_components import make_altair_plot_component, make_space_altair
 from .solara_viz import JupyterViz, SolaraViz
 from .user_param import Slider
 
@@ -20,8 +20,8 @@ __all__ = [
     "Slider",
     "SolaraViz",
     "draw_space",
+    "make_altair_plot_component",
     "make_plot_component",
     "make_space_altair",
-    "make_altair_plot_component",
     "make_space_component",
 ]
