### 2025-01-26 d1c7965
takes care of valueError and CancelledError
takes care of valueError and CancelledError.
pending changes:clean code
