### 2025-01-26 8b8cd19
final changes
optimal performance in browser and jupyter
