+++ mesa/visualization/solara_viz.py
@@ -26,6 +26,7 @@ from __future__ import annotations
 import asyncio
 import inspect
 import threading
+import time
 from collections.abc import Callable
 from typing import TYPE_CHECKING, Literal
 
@@ -37,7 +38,7 @@ from mesa.experimental.devs.simulator import Simulator
 from mesa.mesa_logging import create_module_logger, function_logger
 from mesa.visualization.user_param import Slider
 from mesa.visualization.utils import force_update, update_counter
-import time
+
 if TYPE_CHECKING:
     from mesa.model import Model
 
@@ -234,17 +235,13 @@ def ModelController(
         model_parameters = {}
     model_parameters = solara.use_reactive(model_parameters)
     pause_vis_event = threading.Event()
-    #pause_step_event = threading.Event()
-
+    
     def step():
         try:
             loop = asyncio.new_event_loop()
             asyncio.set_event_loop(loop)
             while running.value and playing.value:
                 time.sleep(play_interval.value / 1000)
-                #if use_threads.value:
-                    #pause_step_event.wait()
-                    #pause_step_event.clear()
                 do_step()
                 if use_threads.value:
                     pause_vis_event.set()
@@ -259,24 +256,15 @@ def ModelController(
             try:
                 loop = asyncio.new_event_loop()
                 asyncio.set_event_loop(loop)
-                #pause_step_event.set()
-                #render_step=-1
                 while playing.value and running.value:
-                    #if render_step!=model.value.steps:
-                    print("before waiting",model.value.steps)
                     pause_vis_event.wait()
                     pause_vis_event.clear()
-                    print("after waiting",model.value.steps)
                     force_update()
-                    print("after rendering",model.value.steps)
-                    #    render_step=model.value.steps
-                    #pause_step_event.set()
             except Exception as e:
                 print(f"Error in vis task: {e}")
             finally:
                 loop.close()
 
-    # h216 result error?
     solara.lab.use_task(step, dependencies=[playing.value, running.value])
 
     solara.use_thread(vis, dependencies=[playing.value])
@@ -286,21 +274,16 @@ def ModelController(
         """Advance the model by the number of steps specified by the render_interval slider."""
         if playing.value:
             for _ in range(render_interval.value):
-                print("model step:",model.value.steps)
                 model.value.step()
                 running.value = model.value.running
                 if not playing.value:
                     break
-
             if not use_threads.value:
-                print("rendering")
                 force_update()
-                print("renderedd")
-
+               
         else:
             for _ in range(render_interval.value):
                 model.value.step()
-                print("model step:",model.value.steps)
                 running.value = model.value.running
             force_update()
 
