### 2025-01-26 1c276aa
Display message for adjusting play interval when using threads
