### 2025-01-26 2fc8eff
Fix: Property layer vizualization for HexGrid
