### 2025-01-26 d0bfcbb
added pre-commit suggestions
