+++ tests/test_solara_viz.py
@@ -9,6 +9,7 @@ import solara
 import mesa
 import mesa.visualization.components.altair_components
 import mesa.visualization.components.matplotlib_components
+from mesa.visualization.components.altair_components import make_altair_space
 from mesa.visualization.components.matplotlib_components import make_mpl_space_component
 from mesa.visualization.solara_viz import (
     Slider,
@@ -132,6 +133,34 @@ def test_call_space_drawer(mocker):  # noqa: D103
     # should call default method with class instance and agent portrayal
     assert mock_space_matplotlib.call_count == 0
     assert mock_space_altair.call_count == 0
+
+    # checking if SpaceAltair is working as intended with post_process
+    def mock_post_process(chart):
+        return chart.configure_legend(titleFontSize=14, labelFontSize=12)
+
+    mock_post_process_spy = mocker.spy(mock_post_process)
+    solara.render(
+        SolaraViz(
+            model,
+            components=[
+                make_altair_space(
+                    agent_portrayal,
+                    propertylayer_portrayal,
+                    post_process=mock_post_process,
+                )
+            ],
+        )
+    )
+
+    mock_space_altair.assert_called_with(
+        model, agent_portrayal, propertylayer_portrayal, post_process=mock_post_process
+    )
+    mock_post_process_spy.assert_called_once()
+    assert mock_space_matplotlib.call_count == 0
+
+    mock_space_altair.reset_mock()
+    mock_space_matplotlib.reset_mock()
+    mock_post_process_spy.reset_mock()
 
     # specify a custom space method
     class AltSpace:
