+++ mesa/visualization/components/altair_components.py
@@ -29,8 +29,8 @@ def make_altair_space(
 
     Args:
         agent_portrayal: Function to portray agents.
-        propertylayer_portrayal: A user specified callable that will be called with the Chart instance from Altair. Allows for fine tuning plots (e.g., control ticks)
-        post_process :not yet implemented
+        propertylayer_portrayal: not yet implemented
+        post_process :A user specified callable that will be called with the Chart instance from Altair. Allows for fine tuning plots (e.g., control ticks)
         space_drawing_kwargs : not yet implemented
 
     ``agent_portrayal`` is called with an agent and should return a dict. Valid fields in this dict are "color",
