### 2025-01-27 a88d816
Update altair_components.py
"alt defined"
