+++ mesa/visualization/components/altair_components.py
@@ -5,7 +5,7 @@ import warnings
 from collections.abc import Callable
 
 import solara
-
+alt=None
 with contextlib.suppress(ImportError):
     import altair as alt
 
