### 2025-01-27 e7f90a3
fixed bugs and added test for post_process
