+++ mesa/visualization/components/altair_components.py
@@ -1,12 +1,9 @@
 """Altair based solara components for visualization mesa spaces."""
 
-import contextlib
 import warnings
 
+import altair as alt
 import solara
-
-with contextlib.suppress(ImportError):
-    import altair as alt
 
 from mesa.experimental.cell_space import DiscreteSpace, Grid
 from mesa.space import ContinuousSpace, _Grid
