### 2025-01-27 #7668 @EwoutH
Thanks for getting back!

I’m starting a new job around then, so I was hoping on a bit earlier. But since Joël will likely review this week or next, we can already work on that feedback.

Keep us in the loop!
