### 2025-01-27 b5c9918
fix: update datacollection validation and add test cases
