### 2025-01-27 95a607b
Update altair_components.py
"removed a line"
