### 2025-01-27 2b69a62
Update altair_components.py
"removed suppress"
