+++ mesa/visualization/components/altair_components.py
@@ -5,7 +5,6 @@ from collections.abc import Callable
 
 import solara
 
-
 from mesa.experimental.cell_space import DiscreteSpace, Grid
 from mesa.space import ContinuousSpace, _Grid
 from mesa.visualization.utils import update_counter
