+++ mesa/visualization/components/altair_components.py
@@ -1,4 +1,5 @@
 """Altair based solara components for visualization mesa spaces."""
+
 import contextlib
 import warnings
 from collections.abc import Callable
@@ -7,7 +8,7 @@ import solara
 
 with contextlib.suppress(ImportError):
     import altair as alt
-    
+
 from mesa.experimental.cell_space import DiscreteSpace, Grid
 from mesa.space import ContinuousSpace, _Grid
 from mesa.visualization.utils import update_counter
