### 2025-01-27 da8fd53
Update altair_components.py
"operator bug"
