### 2025-01-28 efed03e
Fix: Property layer visualization for HexGrid (#2646)
