### 2025-01-28 bfb1c17
Revert "Fix: Property layer visualization for HexGrid (#2646)"
This reverts commit efed03e643fa3269dec33ee492d891782fa9fa61.
