+++ mesa/visualization/mpl_space_drawing.py
@@ -9,8 +9,7 @@ for a paper.
 import contextlib
 import itertools
 import warnings
-from collections.abc import Callable, Iterator
-from functools import lru_cache
+from collections.abc import Callable
 from itertools import pairwise
 from typing import Any
 
@@ -19,7 +18,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
 from matplotlib.cm import ScalarMappable
-from matplotlib.collections import LineCollection, PatchCollection, PolyCollection
+from matplotlib.collections import LineCollection, PatchCollection
 from matplotlib.colors import LinearSegmentedColormap, Normalize, to_rgba
 from matplotlib.patches import Polygon
 
@@ -160,37 +159,6 @@ def draw_space(
     return ax
 
 
-@lru_cache(maxsize=1024, typed=True)
-def _get_hexmesh(
-    width: int, height: int, size: float = 1.0
-) -> Iterator[list[tuple[float, float]]]:
-    """Generate hexagon vertices for the mesh. Yields list of vertex coordinates for each hexagon."""
-
-    # Helper function for getting the vertices of a hexagon given the center and size
-    def _get_hex_vertices(
-        center_x: float, center_y: float, size: float = 1.0
-    ) -> list[tuple[float, float]]:
-        """Get vertices for a hexagon centered at (center_x, center_y)."""
-        vertices = [
-            (center_x, center_y + size),  # top
-            (center_x + size * np.sqrt(3) / 2, center_y + size / 2),  # top right
-            (center_x + size * np.sqrt(3) / 2, center_y - size / 2),  # bottom right
-            (center_x, center_y - size),  # bottom
-            (center_x - size * np.sqrt(3) / 2, center_y - size / 2),  # bottom left
-            (center_x - size * np.sqrt(3) / 2, center_y + size / 2),  # top left
-        ]
-        return vertices
-
-    x_spacing = np.sqrt(3) * size
-    y_spacing = 1.5 * size
-
-    for row, col in itertools.product(range(height), range(width)):
-        # Calculate center position with offset for even rows
-        x = col * x_spacing + (row % 2 == 0) * (x_spacing / 2)
-        y = row * y_spacing
-        yield _get_hex_vertices(x, y, size)
-
-
 def draw_property_layers(
     space, propertylayer_portrayal: dict[str, dict[str, Any]], ax: Axes
 ):
@@ -237,73 +205,45 @@ def draw_property_layers(
         vmax = portrayal.get("vmax", np.max(data))
         colorbar = portrayal.get("colorbar", True)
 
-        # Prepare colormap
+        # Draw the layer
         if "color" in portrayal:
+            data = data.T
             rgba_color = to_rgba(portrayal["color"])
+            normalized_data = (data - vmin) / (vmax - vmin)
+            rgba_data = np.full((*data.shape, 4), rgba_color)
+            rgba_data[..., 3] *= normalized_data * alpha
+            rgba_data = np.clip(rgba_data, 0, 1)
             cmap = LinearSegmentedColormap.from_list(
                 layer_name, [(0, 0, 0, 0), (*rgba_color[:3], alpha)]
             )
+            im = ax.imshow(
+                rgba_data,
+                origin="lower",
+            )
+            if colorbar:
+                norm = Normalize(vmin=vmin, vmax=vmax)
+                sm = ScalarMappable(norm=norm, cmap=cmap)
+                sm.set_array([])
+                ax.figure.colorbar(sm, ax=ax, orientation="vertical")
+
         elif "colormap" in portrayal:
             cmap = portrayal.get("colormap", "viridis")
             if isinstance(cmap, list):
[…]
