### 2025-01-28 4e0dec9
uses array for hex grid property layer fix
