### 2025-01-28 5e5b2cf
solving the 'non rendering of plots while using threads' bug
In use_threads, the threads for step and vis may compete for GI leading to slower execution, especially when sharing resources. In contrast, use_task, which is asyncio-based, avoids GIL contention.
