+++ mesa/visualization/solara_viz.py
@@ -268,9 +268,9 @@ def ModelController(
             finally:
                 loop.close()
 
-    solara.lab.use_task(step, dependencies=[playing.value, running.value])
+    solara.lab.use_task(step, dependencies=[playing.value, running.value],prefer_threaded=True)
 
-    solara.use_thread(vis, dependencies=[playing.value, running.value])
+    solara.lab.use_task(vis, dependencies=[playing.value,running.value],prefer_threaded=True)
 
     @function_logger(__name__)
     def do_step():
@@ -281,12 +281,14 @@ def ModelController(
                 running.value = model.value.running
                 if not playing.value:
                     break
+            if not use_threads.value:
+                force_update()
+
         else:
             for _ in range(render_interval.value):
                 model.value.step()
                 running.value = model.value.running
-
-        force_update()
+            force_update()
 
     @function_logger(__name__)
     def do_reset():
@@ -339,7 +341,6 @@ def SimulatorController(
         play_interval: Interval for playing the model steps in milliseconds.
         render_interval: Controls how often the plots are updated during simulation steps.Higher values reduce update frequency.
         use_threads: Flag for indicating whether to utilize multi-threading for model execution.
-
     Notes:
         The `step button` increments the step by the value specified in the `render_interval` slider.
         This behavior ensures synchronization between simulation steps and plot updates.
@@ -374,8 +375,8 @@ def SimulatorController(
 
     solara.lab.use_task(step, dependencies=[playing.value, running.value])
 
-    solara.use_thread(vis, dependencies=[playing.value, running.value])
-
+    solara.use_thread(vis, dependencies=[playing.value,running.value])
+    
     def do_step():
         """Advance the model by the number of steps specified by the render_interval slider."""
         simulator.run_for(render_interval.value)
