### 2025-01-29 939bd74
Update Binder environment to use latest Mesa version (#2652)
