+++ binder/environment.yml
@@ -1,8 +1,8 @@
-name: example-environment
+name: mesa-tutorials
 channels:
   - conda-forge
 dependencies:
-  - python
+  - python=3.9
   - numpy
   - pip
   - pip:
