### 2025-01-29 31fa87d
Update Binder environment to use latest Mesa version (#2652)
