+++ mesa/visualization/components/altair_components.py
@@ -1,16 +1,19 @@
 """Altair based solara components for visualization mesa spaces."""
 
 import contextlib
+import math
 import warnings
 from collections.abc import Callable
 
 import solara
 
+import mesa
+
 with contextlib.suppress(ImportError):
     import altair as alt
 
-from mesa.experimental.cell_space import DiscreteSpace, Grid
-from mesa.space import ContinuousSpace, _Grid
+from mesa.experimental.cell_space import Grid, HexGrid
+from mesa.space import ContinuousSpace, NetworkGrid, _Grid
 from mesa.visualization.utils import update_counter
 
 
@@ -52,7 +55,6 @@ def make_altair_space(
 
     ``agent_portrayal`` is called with an agent and should return a dict. Valid fields in this dict are "color",
     "size", "marker", and "zorder". Other field are ignored and will result in a user warning.
-
 
     Returns:
         function: A function that creates a SpaceMatplotlib component
@@ -192,90 +194,103 @@ def SpaceAltair(model, agent_portrayal, dependencies: list[any] | None = None):
     solara.FigureAltair(chart)
 
 
-def _get_agent_data_old__discrete_space(space, agent_portrayal):
-    """Format agent portrayal data for old-style discrete spaces.
+def axial_to_pixel(q, r, size=1):
+    """Convert axial coordinates (q, r) to pixel coordinates for hexagonal grid."""
+    x = size * math.sqrt(3) * (q + r / 2)
+    y = size * 1.5 * r
+    return x, y
+
+
+def get_agent_data(space, agent_portrayal):
+    """Generic method to extract agent data for visualization across all space types.
 
     Args:
-        space: the mesa.space._Grid instance
-        agent_portrayal: the agent portrayal callable
+        space: Mesa space object
+        agent_portrayal: Function defining agent visualization properties
 
     Returns:
-        list of dicts
-
-    """
-    all_agent_data = []
-    for content, (x, y) in space.coord_iter():
-        if not content:
-            continue
-        if not hasattr(content, "__iter__"):
-            # Is a single grid
-            content = [content]  # noqa: PLW2901
-        for agent in content:
-            # use all data from agent portrayal, and add x,y coordinates
-            agent_data = agent_portrayal(agent)
-            agent_data["x"] = x
-            agent_data["y"] = y
-            all_agent_data.append(agent_data)
-    return all_agent_data
-
-
-def _get_agent_data_new_discrete_space(space: DiscreteSpace, agent_portrayal):
-    """Format agent portrayal data for new-style discrete spaces.
-
-    Args:
-        space: the mesa.experiment.cell_space.Grid instance
-        agent_portrayal: the agent portrayal callable
-
-    Returns:
-        list of dicts
-
+        List of agent data dictionaries with coordinates
     """
     all_agent_data = []
 
-    for cell in space.all_cells:
-        for agent in cell.agents:
-            agent_data = agent_portrayal(agent)
-            agent_data["x"] = cell.coordinate[0]
-            agent_data["y"] = cell.coordinate[1]
-            all_agent_data.append(agent_data)
-    return all_agent_data
+    # New DiscreteSpace
+    if isinstance(space, Grid):
+        for cell in space.all_cells:
+            for agent in cell.agents:
+                data = agent_portrayal(agent)
+                data.update({"x": cell.coordinate[0], "y": cell.coordinate[1]})
+                all_agent_data.append(data)
 
+    # Legacy Grid
+    elif isinstance(space, _Grid):
+        for content, (x, y) in space.coord_iter():
+            if not content:
+                continue
+            agents = [content] if not hasattr(content, "__iter__") else content
+            for agent in agents:
+                data = agent_portrayal(agent)
[…]
