### 2025-01-31 569916f
Remove pinned Python version and ensure compatibility with Python >=3.11
