+++ pyproject.toml
@@ -53,7 +53,7 @@ network = [
 viz = [
   "matplotlib",
   "solara",
-  "altair"
+  "altair",
 ]
 # Dev and CI stuff
 dev = [
