+++ mesa/visualization/components/altair_components.py
@@ -2,7 +2,6 @@
 
 import contextlib
 import math
-
 import warnings
 from collections.abc import Callable
 
@@ -17,7 +16,6 @@ with contextlib.suppress(ImportError):
 
 from mesa.experimental.cell_space import Grid
 from mesa.space import ContinuousSpace, NetworkGrid, _Grid
-
 from mesa.visualization.utils import update_counter
 
 
@@ -104,7 +102,6 @@ def make_altair_plot_component(
 
 
 @solara.component
-
 def PlotAltair(
     model,
     measure: str | dict[str, str] | list[str] | tuple[str],
