### 2025-02-01 2740d3f
Refactored draw_space to separate grid, properties, and agents (#2638)
