### 2025-02-01 3ba3593
Change from use_thread to use_task in SimulatorController for better control
This should improve the overall functionality and reliability of the SimulatorController, particularly in scenarios where stopping the execution is necessary.
