### 2025-02-02 822d3dc
Update docs/getting_started.md
