+++ mesa/visualization/__init__.py
@@ -1,7 +1,9 @@
-"""Solara based visualization for Mesa models."""
+"""Solara based visualization for Mesa models.
 
 .. note::
     SolaraViz is experimental and still in active development in Mesa 3.x. While we attempt to minimize them, there might be API breaking changes in minor releases.
+
+"""
 
 from mesa.visualization.mpl_space_drawing import (
     draw_space,
