### 2025-02-04 a1089ad
Feat: Added portrayal components for agent and property layer
