+++ mesa/visualization/components/portrayal_components.py
@@ -1,0 +1,159 @@
+from dataclasses import dataclass, field, asdict
+from typing import Optional, Union, ClassVar, Set, Dict
+from matplotlib import pyplot as plt
+import matplotlib.colors as mcolors
+from numbers import Number
+from matplotlib.colors import Colormap
+
+@dataclass
+class AgentPortrayalStyle:
+    """
+    A class to define the visual style of an agent in Mesa visualizations.
+
+    Attributes:
+        color: Color specification for the agent. Can be any valid matplotlib color.
+        marker: Marker style for the agent (e.g., 'o', 's', '^').
+        size: Size of the agent marker in points.
+        zorder: The drawing order of the agent relative to other elements.
+
+    Example:
+        >>> def agent_portrayal(agent):
+        >>>     portrayal = AgentPortrayalStyle(
+        >>>         color='tab:orange' if agent.id == 1 else 'tab:blue',
+        >>>         marker='^',
+        >>>         size=70,
+        >>>         zorder=1
+        >>>     )
+        >>>     return dict(portrayal)       
+    """
+    
+    VALID_MARKERS: ClassVar[Set[str]] = {
+        'o', 's', '^', 'v', '<', '>', 'D', 'p', 'h', 'H', '8', '*', 
+        '+', 'x', '.', ',', '1', '2', '3', '4', '|', '_',
+    }
+
+    # Default values
+    color: Optional[Union[str, tuple]] = "tab:blue"
+    marker: Optional[str] = "o"
+    size: Optional[Number] = 50
+    zorder: Optional[int] = 1
+
+    def __post_init__(self):
+        """Validate the attributes after initialization."""
+        if self.color is not None:
+            try:
+                mcolors.to_rgb(self.color)
+            except ValueError:
+                raise ValueError(f"Invalid color specification: {self.color}")
+        
+        if self.marker is not None:
+            if not isinstance(self.marker, str):
+                raise ValueError(f"Marker must be a string, got {type(self.marker)}")
+            if self.marker not in self.VALID_MARKERS:
+                raise ValueError(
+                    f"Invalid marker '{self.marker}'. Valid markers are: {', '.join(sorted(self.VALID_MARKERS))}"
+                )
+        
+        if self.size is not None and not isinstance(self.size, Number):
+            raise ValueError(f"Size must be a number, got {type(self.size)}")
+        
+        if self.zorder is not None and not isinstance(self.zorder, int):
+            raise ValueError(f"Zorder must be an integer, got {type(self.zorder)}")
+        
+    def __iter__(self):
+        return iter(asdict(self).items())
+
+
+@dataclass
+class PropertyLayerStyle:
+    """Style configuration for a single property layer.
+        
+    Args:
+        color: A valid matplotlib color string
+        colormap: A valid matplotlib colormap name or Colormap object
+        vmin: Minimum value for colormap scaling
+        vmax: Maximum value for colormap scaling
+        alpha: Opacity value between 0 and 1
+        colorbar: Whether to show colorbar
+
+    Note: color and colormap are mutually exclusive.
+    """
+    vmin: Optional[float] = None
+    vmax: Optional[float] = None
+    alpha: Optional[float] = None
+    colorbar: Optional[bool] = None
+    color: Optional[str] = None
+    colormap: Optional[Union[str, Colormap]] = None
+    
+    def __post_init__(self):
+        """Validate style attributes."""
+        # Validate color/colormap exclusivity
+        if self.color is not None and self.colormap is not None:
+            raise ValueError("Cannot specify both color and colormap")
+        if self.color is not None and self.colormap is not None:
+            raise ValueError("Cannot specify both color and colormap")
+            
+        # Validate color if specified
[…]
