### 2025-02-04 17a1a9b
Doc: added necessary docstrings
