+++ mesa/visualization/components/portrayal_components.py
@@ -1,14 +1,22 @@
-from dataclasses import dataclass, field, asdict
-from typing import Optional, Union, ClassVar, Set, Dict
-from matplotlib import pyplot as plt
-import matplotlib.colors as mcolors
+"""Components for defining visual styles and portrayals in Mesa visualizations.
+
+This module provides classes for configuring the visual appearance of agents and
+property layers in Mesa visualization components.
+"""
+
+from collections.abc import Iterator
+from dataclasses import asdict, dataclass, field
 from numbers import Number
+from typing import ClassVar
+
+import matplotlib.colors as mcolors
+from matplotlib import pyplot as plt
 from matplotlib.colors import Colormap
+
 
 @dataclass
 class AgentPortrayalStyle:
-    """
-    A class to define the visual style of an agent in Mesa visualizations.
+    """A class to define the visual style of an agent in Mesa visualizations.
 
     Attributes:
         color: Color specification for the agent. Can be any valid matplotlib color.
@@ -24,28 +32,30 @@ class AgentPortrayalStyle:
         >>>         size=70,
         >>>         zorder=1
         >>>     )
-        >>>     return dict(portrayal)       
+        >>>     return dict(portrayal)
     """
-    
-    VALID_MARKERS: ClassVar[Set[str]] = {
-        'o', 's', '^', 'v', '<', '>', 'D', 'p', 'h', 'H', '8', '*', 
-        '+', 'x', '.', ',', '1', '2', '3', '4', '|', '_',
+
+    # fmt:off
+    VALID_MARKERS: ClassVar[set[str]] = {
+        "o", "s", "^", "v", "<", ">", "D", "p", "h", "H", "8", "*",
+        "+", "x", ".", ",", "1", "2", "3", "4", "|", "_",
     }
+    # fmt:on
 
     # Default values
-    color: Optional[Union[str, tuple]] = "tab:blue"
-    marker: Optional[str] = "o"
-    size: Optional[Number] = 50
-    zorder: Optional[int] = 1
+    color: str | tuple | None = "tab:blue"
+    marker: str | None = "o"
+    size: Number | None = 50
+    zorder: int | None = 1
 
     def __post_init__(self):
         """Validate the attributes after initialization."""
         if self.color is not None:
             try:
                 mcolors.to_rgb(self.color)
-            except ValueError:
-                raise ValueError(f"Invalid color specification: {self.color}")
-        
+            except ValueError as err:
+                raise ValueError(f"Invalid color specification: {self.color}") from err
+
         if self.marker is not None:
             if not isinstance(self.marker, str):
                 raise ValueError(f"Marker must be a string, got {type(self.marker)}")
@@ -53,21 +63,22 @@ class AgentPortrayalStyle:
                 raise ValueError(
                     f"Invalid marker '{self.marker}'. Valid markers are: {', '.join(sorted(self.VALID_MARKERS))}"
                 )
-        
+
         if self.size is not None and not isinstance(self.size, Number):
             raise ValueError(f"Size must be a number, got {type(self.size)}")
-        
+
         if self.zorder is not None and not isinstance(self.zorder, int):
             raise ValueError(f"Zorder must be an integer, got {type(self.zorder)}")
-        
-    def __iter__(self):
+
+    def __iter__(self) -> Iterator[tuple[str, any]]:
+        """Return an iterator of the style attributes and their values."""
         return iter(asdict(self).items())
 
 
 @dataclass
 class PropertyLayerStyle:
     """Style configuration for a single property layer.
-        
+
     Args:
         color: A valid matplotlib color string
         colormap: A valid matplotlib colormap name or Colormap object
[…]
