### 2025-02-04 9af4323
fix: added null checks
