### 2025-02-06 f50fefb
test: improve test coverage for DataCollector validation
