### 2025-02-06 615a096
test: improve test coverage for DataCollector validation
