### 2025-02-06 #7668 @jofmi
Dear All, I have now finished my review, thanks for the patience!

First of all, congratulations on this new version of MESA! The package has improved significantly since I last worked with it many years ago. Many of the challenges I encountered back then have been addressed. In particular, I think it was a good decision to deprecate the scheduler and the new visualization module is very impressive.

For transparency, I have previously also developed a Python-based ABM package (AgentPy). Given the advancements in this new version of MESA, I will also update the AgentPy documentation to reflect that the package is now largely deprecated in favor of MESA.

Below, you will find my comments presented in a checklist format. This will allow us to systematically address and mark them as resolved.

Major comments
- [ ] Experimental features: cell spaces and event-based scheduling (lines 76 & 86)
	- I believe these  features should be completed before publication. Publishing a work in progress is not ideal, particularly since this is not the first paper on this package and these features appear to be close to a stable release.
	- Additionally, when I attempted to test these new features - which I find to be very interesting innovations - I found the online documentation lacking sufficient information.
	- If completing these features before publication is not feasible, I recommend moving them to a dedicated section such as _Future Plans_ or _Next Steps_.

Comments on the documentation
- [ ] On the index page, I would add a sentence explaining what the "rec" and "all" extensions are needed for respectively (e.g. the visualization package?).
- [ ] I would suggest to put the "Overview of the MESA library" section within "Getting started" as a seperate page, as it is a handy reference that is more accessible than the API documentation, also after getting started.
- [ ] In addition, it would be great for the overview to cover all features of MESA. Compared to the paper, spaces and time advancement are missing here.
- [ ] In the introductory tutorial, it is a bit confusing what is supposed to go in the jupyter or the python file. For simplification of the tutorial, I would recommend for the whole tutorial to be within the jupyter file. If you want to keep the recommendation to split up files, I would add it as a tipp in "Best practices".
- [ ] For the setup, using code cells with terminal commands "!pip install --upgrade mesa[rec]" (note the exclamation mark at the start) enables installation within the jupyter file, removing the need for special instructions for google collab.
- [ ] Please put line breaks into the comments that are longer than 79 characters, allowing the documentation do be read on half-screen.
- [ ] Some sentences do not start capitalized.
- [ ] In the visualization tutorial, I would put the whole code of "MoneyModel.py" within the jupyter notebook, removing the need to search through the first tutorial for a file that is not really described.
- [ ] Maybe mention that the visualization needs the "rec" dependencies in the installation?
- [ ] I find the "Choose version" button at the top confusing, as it does not display the currently chosen version and there is already an indication on the bottom right.

Comments on the code
- [ ] The quality of the code and tests seem very good!
- [ ] Typing
	- One thing that could be positively highlighted is that MESA is using type hints. However, some additions would be needed for it to be able to run in a strongly typed environment.
	- In the introductory tutorial, chapter "Agents Exchange"/"Running your first model", I get a type error for "other_agent.wealth"/"a.wealth". It seems that model.agents does not return the correct agent type.
	- Same with using self.model within agent, the Agent does not know about the model types. (e.g. model.grid)

Comments on the paper
- [ ] l16 The Statement of Need could be more clearly structured. Currently, it reads more like a historical recap. I recommend organizing it along these lines: (1) Why such a package is needed, (2) The current state of the field, and (3) What MESA contributes — focusing on its present state rather than its 2014 origins. Additionally, please refer to the _Statement of Need_ and _State of the Field_ requirements outlined in the JOSS checklist above to ensure alignment with the journal's expectations.
- [ ] l41 For consistency with the other sections, please add a code example of instantiating a model and passing a seed
- [ ] l46 Similarly, please add a code example of instantiating and subclassing an agent
- [ ] l80 & l86 Maybe call these two 'modes' "discrete time" and "continuous time", if I understood it correctly? I think that would make it easier to understand.
- [ ] l109 I would suggest to move the "Applications" section to the start of the paper
- [ ] l134 Acknowledgements are usually placed after the conclusions

Smaller comments on the paper
- [ ] l15 Change to 2025
- [ ] l17 Not all ABMs represent artificial societies, here it is an implied synonym
- [ ] l20 The two references need to be placed into the same bracket
- [ ] l28 I think it should say here: "an" improved visualization framework
- [ ] l34 What does pure python mean? Does it not also use C, JS/React, etc.?
- [ ] l35 Markdown mistake in list formatting
- [ ] l43 Word repetition: "instantiates"
- [ ] l46 Please add a sentence describing agent types, to understand what the types mean later at "Data collection"
- [ ] l62-64 Please shortly explain these terms or add a reference
- [ ] l94 Markdown mistake, close bracket for link
- [ ] l96 Please add a link or reference regarding the Solara package
- [ ] l97 Please use a code example and screenshot that correspond to each other, demonstrating what kind of code creates what kind of visualization.
- [ ] l98 Maybe add a sentence how this visualization module aligns or differs with the established standard of NetLogo.
- [ ] l126 Markdown mistake in list formatting
- [ ] l133 Maybe add a sentence on which range of topics are covered by these community extensions and add a link or reference to a list or overview of these extensions and tutorials same as with the example models
- [ ] l143 Maybe say "want" instead of "need"?
