### 2025-02-06 projectmesa/mesa#2666 @EwoutH
@jofmi thanks a lot for your detailed and thorough review! It’s really nice to have a fresh sets of eyes on it.

Jackie thanks  for creating the tracking issue.

At some point we had experimental split out in a separate section, maybe we could revert to that. Also, we’re stabilizing the cell space so that shouldn’t be an issue. I see a few options for the other features:
    1. We could split them out in a separate section (like earlier).
    2. We clearly mark all experimental features with an _experimental_ label, in the current structure.

Curious what everybody (including @jofmi) thinks about both.

As on the other feedback:
- Tutorials: @tpike3 since you worked on them recently, would you like to pick this up?
- @quaquel would you like to work on typing?
- I can take some of the documentation issues
- @jackiekazil could you take on the comments on the paper itself?
