### 2025-02-06 #7668 @EwoutH
@jofmi thanks a lot for your detailed and thorough review! It’s really nice to have a fresh sets of eyes on it.

We’re tracking in:
- https://github.com/projectmesa/mesa/issues/2666
