### 2025-02-07 7c77860
Enhance DataCollector to validate model_reporters functions (#2605)
