### 2025-02-09 f0da710
Feat: Added command center
