### 2025-02-09 140cff1
docs: Clarify recommended dependencies in installation guide
Add explanation of what the [rec] optional dependencies provide, including
visualization (SolaraViz), Matplotlib plotting, and network modeling capabilities.
This helps users understand when they need to install these additional packages.
