### 2025-02-09 a2e38b9
docs: Split off the overview to a separate page
Based on feedback from the paper, this splits the overview (again) off to a separate page. It does add one more item to the top bar, but I think that's worth it.
