+++ docs/overview.md
@@ -1,0 +1,214 @@
+## Overview of the MESA library
+
+Mesa is modular, meaning that its modeling, analysis and visualization components are kept separate but intended to work together. The modules are grouped into three categories:
+
+1. **Modeling:** Classes used to build the models themselves: a model and agent classes, space for them to move around in, and built-in functionality for managing agents.
+2. **Analysis:** Tools to collect data generated from your model, or to run it multiple times with different parameter values.
+3. **Visualization:** Classes to create and launch an interactive model visualization, using a browser-based interface.
+
+### Modeling modules
+
+Most models consist of one class to represent the model itself and one or more classes for agents. Mesa provides built-in functionality for managing agents and their interactions. These are implemented in Mesa's modeling modules:
+
+- [mesa.model](apis/model)
+- [mesa.agent](apis/agent)
+- [mesa.space](apis/space)
+
+The skeleton of a model might look like this:
+
+```python
+import mesa
+
+class MyAgent(mesa.Agent):
+    def __init__(self, model, age):
+        super().__init__(model)
+        self.age = age
+
+    def step(self):
+        self.age += 1
+        print(f"Agent {self.unique_id} now is {self.age} years old")
+        # Whatever else the agent does when activated
+
+class MyModel(mesa.Model):
+    def __init__(self, n_agents):
+        super().__init__()
+        self.grid = mesa.space.MultiGrid(10, 10, torus=True)
+        for _ in range(n_agents):
+            initial_age = self.random.randint(0, 80)
+            a = MyAgent(self, initial_age)
+            coords = (self.random.randrange(0, 10), self.random.randrange(0, 10))
+            self.grid.place_agent(a, coords)
+
+    def step(self):
+        self.agents.shuffle_do("step")
+```
+
+If you instantiate a model and run it for one step, like so:
+
+```python
+model = MyModel(5)
+model.step()
+```
+
+You should see agents 1-5, activated in random order. See the [tutorial](tutorials/intro_tutorial) or API documentation for more detail on how to add model functionality.
+
+
+### AgentSet and model.agents
+Mesa 3.0 makes `model.agents` and the AgentSet class central in managing and activating agents.
+
+#### model.agents
+`model.agents` is an AgentSet containing all agents in the model. It's automatically updated when agents are added or removed:
+
+```python
+# Get total number of agents
+num_agents = len(model.agents)
+
+# Iterate over all agents
+for agent in model.agents:
+    print(agent.unique_id)
+```
+
+#### AgentSet Functionality
+AgentSet offers several methods for efficient agent management:
+
+1. **Selecting**: Filter agents based on criteria.
+   ```python
+   high_energy_agents = model.agents.select(lambda a: a.energy > 50)
+   ```
+2. **Shuffling and Sorting**: Randomize or order agents.
+   ```python
+   shuffled_agents = model.agents.shuffle()
+   sorted_agents = model.agents.sort(key="energy", ascending=False)
+   ```
+3. **Applying methods**: Execute methods on all agents.
+   ```python
+   model.agents.do("step")
+   model.agents.shuffle_do("move")  # Shuffle then apply method
+   ```
+4. **Aggregating**: Compute aggregate values across agents.
+   ```python
+   avg_energy = model.agents.agg("energy", func=np.mean)
+   ```
+5. **Grouping**: Group agents by attributes.
+   ```python
+   grouped_agents = model.agents.groupby("species")
+
+   for _, agent_group in grouped_agents:
+      agent_group.shuffle_do()
+   species_counts = grouped_agents.count()
+   mean_age_by_group = grouped_agents.agg("age", np.mean)
+   ```
+`model.agents` can also be accessed within a model instance using `self.agents`.
+
[…]
