### 2025-02-09 f59fa99
docs: Add space and time advancement sections to overview
Expand the Mesa overview documentation to fully cover all core features:
- Add comprehensive section on spaces including grid-based, network, and
  Voronoi spaces, with code examples
- Add section on property layers and continuous space functionality
- Add new section on time advancement covering basic steps, agent
  activation patterns, and event-based scheduling
- Include practical code examples demonstrating each feature
- Maintain consistency with existing documentation style and structure

This update brings the overview in line wi
[…]
