### 2025-02-09 55767e9
index: remove specific libraries
