### 2025-02-10 df7d656
feat: exec() vulnerablility checking
