### 2025-02-10 c84aacc
docs: Clarify recommended dependencies in installation guide (#2672)
Add explanation of what the [rec] optional dependencies provide, including
visualization (SolaraViz), Matplotlib plotting, and network modeling capabilities.
This helps users understand when they need to install these additional packages.
