### 2025-02-10 09cfde8
docs: Split off the overview page, extend with spaces/activation (#2673)
* docs: Split off the overview to a separate page

Based on feedback from the paper, this splits the overview (again) off to a separate page. It does add one more item to the top bar, but I think that's worth it.

* docs: Add space and time advancement sections to overview

Expand the Mesa overview documentation to fully cover all core features:
- Add comprehensive section on spaces including grid-based, network, and
  Voronoi spaces, with code examples
- Add section on property layers and continuous space functionality
-
[…]
