+++ mesa/examples/basic/conways_game_of_life/agents.py
@@ -1,18 +1,25 @@
 from mesa import Agent
-from mesa.discrete_space import CellAgent
+from mesa.discrete_space import FixedAgent
 
 
-class Cell(CellAgent):
+class Cell(FixedAgent):
     """Represents a single ALIVE or DEAD cell in the simulation."""
 
     DEAD = 0
     ALIVE = 1
 
+    @property
+    def x(self):
+        return self.cell.coordinate[0]
+
+    @property
+    def y(self):
+        return self.cell.coordinate[1]
+
     def __init__(self, model, cell, init_state=DEAD):
         """Create a cell, in the given state, at the given x, y position."""
         super().__init__(model)
         self.cell = cell
-        self.x, self.y = pos
         self.state = init_state
         self._next_state = None
 
@@ -22,7 +29,7 @@ class Cell(CellAgent):
 
     @property
     def neighbors(self):
-        return self.model.grid.iter_neighbors((self.x, self.y), True)
+        return self.cell.neighborhood.agents
 
     def determine_state(self):
         """Compute if the cell will be dead or alive at the next tick.  This is
