### 2025-02-11 13cbba9
Move game of life to use cell space
