+++ mesa/examples/basic/conways_game_of_life/agents.py
@@ -1,4 +1,3 @@
-from mesa import Agent
 from mesa.discrete_space import FixedAgent
 
 
