+++ mesa/__init__.py
@@ -6,6 +6,7 @@ Core Objects: Model, and Agent.
 import datetime
 
 import mesa.experimental as experimental
+import mesa.discrete_space as discrete_space
 import mesa.space as space
 from mesa.agent import Agent
 from mesa.batchrunner import batch_run
@@ -18,6 +19,7 @@ __all__ = [
     "Model",
     "batch_run",
     "experimental",
+    "discrete_space",
     "space",
 ]
 
