### 2025-02-12 89e1bde
Move Schelling to discrete space
