+++ mesa/examples/basic/schelling/agents.py
@@ -1,23 +1,24 @@
-from mesa import Agent
+from mesa.discrete_space import CellAgent
 
 
-class SchellingAgent(Agent):
+class SchellingAgent(CellAgent):
     """Schelling segregation agent."""
 
-    def __init__(self, model, agent_type: int) -> None:
+    def __init__(self, model, cell, agent_type: int, homophily:float=0.4, radius:int=1) -> None:
         """Create a new Schelling agent.
         Args:
             model: The model instance the agent belongs to
             agent_type: Indicator for the agent's type (minority=1, majority=0)
         """
         super().__init__(model)
+        self.cell = cell
         self.type = agent_type
+        self.homophily = homophily
+        self.radius = radius
 
     def step(self) -> None:
         """Determine if agent is happy and move if necessary."""
-        neighbors = self.model.grid.get_neighbors(
-            self.pos, moore=True, radius=self.model.radius
-        )
+        neighbors = self.cell.get_neighborhood(radius=self.radius).agents
 
         # Count similar neighbors
         similar_neighbors = len([n for n in neighbors if n.type == self.type])
@@ -30,7 +31,7 @@ class SchellingAgent(Agent):
             similarity_fraction = 0.0
 
         # Move if unhappy
-        if similarity_fraction < self.model.homophily:
-            self.model.grid.move_to_empty(self)
+        if similarity_fraction < self.homophily:
+            self.cell = self.model.grid.select_random_empty_cell()
         else:
             self.model.happy += 1
