+++ mesa/examples/basic/schelling/agents.py
@@ -20,7 +20,7 @@ class SchellingAgent(CellAgent):
 
     def step(self) -> None:
         """Determine if agent is happy and move if necessary."""
-        neighbors = self.cell.get_neighborhood(radius=self.radius).agents
+        neighbors = list(self.cell.get_neighborhood(radius=self.radius).agents)
 
         # Count similar neighbors
         similar_neighbors = len([n for n in neighbors if n.type == self.type])
