+++ mesa/__init__.py
@@ -5,9 +5,9 @@ Core Objects: Model, and Agent.
 
 import datetime
 
+import mesa.discrete_space as discrete_space
 import mesa.experimental as experimental
 import mesa.space as space
-import mesa.discrete_space as discrete_space
 from mesa.agent import Agent
 from mesa.batchrunner import batch_run
 from mesa.datacollection import DataCollector
@@ -16,8 +16,7 @@ from mesa.model import Model
 __all__ = [
     "Agent",
     "DataCollector",
-    "discrete_space"
-    "Model",
+    "discrete_spaceModel",
     "batch_run",
     "experimental",
     "space",
