### 2025-02-13 a9fb28a
Move virus_on_a_network to discrete_space
