### 2025-02-14 #7668 @jackiekazil
RE: l62-64 Please shortly explain these terms or add a reference
^^ Do you want all terms explained? Or is there something in particular that needs clarification? Asking for clarification to make sure we hit the mark.
