+++ paper/paper.bib
@@ -119,14 +119,15 @@
 }
 
 @article{taberna2023uncertainty,
-  title={Putting renewable energy auctions into action--An agent-based model of onshore wind power auctions in Germany},
-  author={Anatolitis, Vasilios and Welisch, Marijke},
-  journal={Energy Policy},
-  volume={110},
-  pages={394--402},
-  year={2017},
-  publisher={Elsevier},
-  doi = {10.1016/j.enpol.2017.08.024}
+  title={Uncertainty in boundedly rational household adaptation to environmental shocks},
+  author={Taberna, Alessandro and Filatova, Tatiana and Hadjimichael, Antonia and Noll, Brayton},
+  journal={Proceedings of the National Academy of Sciences},
+  volume={120},
+  number={44},
+  pages={e2215675120},
+  year={2023},
+  publisher={National Acad Sciences},
+  doi = {10.1073/pnas.2215675120}
 }
 
 @article{ghanem2022balancing,
