### 2025-02-14 #7668 @jofmi
@jackiekazil particularly "voronoiMesh for irregular tessellations", but for new readers it might also be good to have a reference or a sentence regarding the moore and vonNeumann difference.

but i won't insist on this point, just thought it could be useful :)
