### 2025-02-14 c937ddf
update joss paper to address some requested edits (#2678)
