### 2025-02-15 projectmesa/mesa#2666 @jackiekazil
@wang-boyu, for the stuff that was merged in #2678, can you mark it off in the list above, so we know  it is complete?
