### 2025-02-15 projectmesa/mesa#2666 @wang-boyu
Done!
