### 2025-02-15 2802cc3
fix multiple citations in joss paper
