### 2025-02-15 fb5b53e
Fix draw_network: Visualizing Network Models with More Agents than Nodes
