+++ mesa/visualization/mpl_space_drawing.py
@@ -483,17 +483,42 @@ def draw_network(
     x_padding = width / 20
     y_padding = height / 20
 
+    node_to_agents = {}
+    for agent in space.agents:  # ✅ Directly access the agent list
+        if hasattr(agent, "pos"):  # Ensure agent has a position
+            if agent.pos not in node_to_agents:
+                node_to_agents[agent.pos] = []
+            node_to_agents[agent.pos].append(agent)
+
+    # Adjust positions for multiple agents per node
+    adjusted_positions = []
+    for node in space.G.nodes: 
+        agents = node_to_agents[node] if node in node_to_agents else []
+        base_x, base_y = pos[node]  
+        num_agents = len(agents)
+
+        for i, agent in enumerate(agents):
+            angle = (2 * np.pi * i) / max(1, num_agents)  
+            offset_x = 0.05 * np.cos(angle)  
+            offset_y = 0.05 * np.sin(angle)  
+            adjusted_positions.append((agent, (base_x + offset_x, base_y + offset_y)))
+    
     # gather agent data
     s_default = (180 / max(width, height)) ** 2
     arguments = collect_agent_data(space, agent_portrayal, size=s_default)
 
-    # this assumes that nodes are identified by an integer
-    # which is true for default nx graphs but might user changeable
-    pos = np.asarray(list(pos.values()))
-    arguments["loc"] = pos[arguments["loc"]]
+    # Update agent locations
+    if len(adjusted_positions) > 0:
+        arguments["loc"] = np.array([pos for _, pos in adjusted_positions], dtype=np.float64).reshape(-1, 2)
+    else:
+        arguments["loc"] = np.zeros((0, 2))  # ✅ Ensure correct shape even when empty
+
+
+
 
     # plot the agents
-    _scatter(ax, arguments, **kwargs)
+    if arguments["loc"].shape[0] > 0:
+        _scatter(ax, arguments, **kwargs)
 
     # further styling
     ax.set_axis_off()
