### 2025-02-15 f777bba
Fix num_agents issue and ensure compatibility with NetworkGrid
