### 2025-02-15 e5579ad
ensuring event objects dont reset during re-renders and removing async tasks
