### 2025-02-15 9a94b2b
docs: Add Roles section to CONTRIBUTING.md
This PR adds a new "Roles" section to the Contributing Guide to clarify Mesa's community structure and progression path. It outlines the different ways community members can contribute and grow within the project:

- Describes the Contributor, Collaborator, Maintainer and Special roles
- Emphasizes that Mesa welcomes contributions from everyone 
- Explains GitHub permissions and labels for each role
- Encourages reaching out to discuss taking on more responsibility
- Reinforces that we're all volunteers working together

This documentation aims to mak
[…]
