+++ mesa/visualization/components/altair_components.py
@@ -63,7 +63,7 @@ def make_altair_space(
     "size", "marker", and "zorder". Other field are ignored and will result in a user warning.
 
     Returns:
-        function: A function that creates a SpaceMatplotlib component
+        function: A function that creates a SpaceAltair component
     """
     if agent_portrayal is None:
 
@@ -113,7 +113,21 @@ def PlotAltair(
     width: int = 500,
     height: int = 300,
 ) -> solara.FigureAltair:
-    """Create an Altair plot for model."""
+    """Create an Altair plot for model data.
+
+    Args:
+        model: The mesa.Model instance.
+        measure (str | dict[str, str] | list[str] | tuple[str]): Measure(s) to plot.
+        post_process: An optional callable that takes an Altair Chart object as
+                      input and returns a modified Chart object. This allows
+                      for customization of the plot (e.g., adding annotations,
+                      changing axis labels).
+        width: The width of the chart in pixels.
+        height: The height of the chart in pixels.
+
+    Returns:
+        A solara.FigureAltair component displaying the generated Altair chart."""
+
     update_counter.get()
     df = model.datacollector.get_model_vars_dataframe().reset_index()
 
@@ -189,8 +203,30 @@ def SpaceAltair(
 ):
     """Create an Altair-based space visualization component.
 
+    Args:
+        model: The mesa.Model instance containing the space to visualize.
+               The model must have a `grid` or `space` attribute that
+               represents the space (e.g., Grid, ContinuousSpace, NetworkGrid).
+        agent_portrayal: A callable that takes an agent as input and returns
+                         a dictionary specifying how the agent should be
+                         visualized.  The dictionary can contain the following keys:
+                           - "color":  A string representing the agent's color (e.g., "red", "#FF0000").
+                           - "size":   A number representing the agent's size.
+                           - "tooltip": A string to display as a tooltip when hovering over the agent.
+                           - Any other Vega-Lite mark properties that are supported by Altair.
+        dependencies: A list of dependencies that trigger a re-render of the
+                      component when they change.  This can be used to update
+                      the visualization when the model state changes.
+        post_process: An optional callable that takes an Altair Chart object
+                      as input and returns a modified Chart object. This allows
+                      for customization of the plot (e.g., adding annotations,
+                      changing axis labels).
+
     Returns:
-        a solara FigureAltair instance
+        A solara.FigureAltair instance, which is a Solara component that
+        renders the Altair chart.
+
+
     """
     update_counter.get()
     space = getattr(model, "grid", None)
@@ -214,11 +250,11 @@ def axial_to_pixel(q, r, size=1):
 
 
 def get_agent_data(space, agent_portrayal):
-    """Draw a Matplotlib-based visualization of the space.
+    """Generic method to extract agent data for visualization across all space types.
 
     Args:
-        space: the space of the mesa model
-        agent_portrayal: A callable that returns a dict specifying how to show the agent
+        space: Mesa space object
+        agent_portrayal: Function defining agent visualization properties
 
     Returns:
         List of agent data dictionaries with coordinates
@@ -235,7 +271,7 @@ def get_agent_data(space, agent_portrayal):
                     all_agent_data.append(data)
 
         case _Grid():
[…]
