+++ mesa/visualization/components/altair_components.py
@@ -336,10 +336,12 @@ def _draw_grid(space, agent_portrayal):
     """
     all_agent_data = get_agent_data(space, agent_portrayal)
 
+
     # Handle empty state
     if not all_agent_data:
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
 
+    
     match space:
         case Grid():
             return _draw_discrete_grid(space, all_agent_data, agent_portrayal)
@@ -389,6 +391,11 @@ def _draw_discrete_grid(space, all_agent_data, agent_portrayal):
         length = min(space.width, space.height)
         chart = chart.mark_point(size=30000 / length**2, filled=True)
 
+    chart = chart.encode(
+        x=alt.X("x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))),
+        y=alt.Y("y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1)))
+    )
+
     return chart
 
 
@@ -426,7 +433,12 @@ def _draw_legacy_grid(space, all_agent_data, agent_portrayal):
         length = min(space.width, space.height)
         chart = chart.mark_point(size=30000 / length**2, filled=True)
 
-    return chart
+    chart = chart.encode(
+        x=alt.X("x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))),
+        y=alt.Y("y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1)))
+    )
+
+    return chart 
 
 
 def _draw_hex_grid(space, all_agent_data, agent_portrayal):
@@ -447,9 +459,23 @@ def _draw_hex_grid(space, all_agent_data, agent_portrayal):
         agent_data["x"] = x
         agent_data["y"] = y
 
+    # Calculate proper bounds that account for the full hexagon width and height
+    x_max = space.width * x_spacing + (space.height % 2) * (x_spacing / 2)
+    y_max = space.height * y_spacing
+
+    # Add padding that accounts for the hexagon points
+    x_padding = size * math.sqrt(3) / 2
+    y_padding = size
+
+    x_scale = alt.Scale(domain=(-2 * x_padding, x_max + x_padding))
+    y_scale = alt.Scale(domain=(-2 * y_padding, y_max + y_padding))
+
+
+
+
     encoding_dict = {
-        "x": alt.X("x", axis=None, type=x_y_type),
-        "y": alt.Y("y", axis=None, type=x_y_type),
+        "x": alt.X("x", axis=None, type=x_y_type,scale=x_scale),
+        "y": alt.Y("y", axis=None, type=x_y_type,scale=y_scale),
         "tooltip": [
             alt.Tooltip(key, type=alt.utils.infer_vegalite_type_for_pandas([value]))
             for key, value in all_agent_data[0].items()
@@ -470,19 +496,6 @@ def _draw_hex_grid(space, all_agent_data, agent_portrayal):
         )
         .mark_point(filled=True)
         .properties(width=280, height=280)
-    )
-
-    # Calculate proper bounds that account for the full hexagon width and height
-    x_max = space.width * x_spacing + (space.height % 2) * (x_spacing / 2)
-    y_max = space.height * y_spacing
-
-    # Add padding that accounts for the hexagon points
-    x_padding = size * math.sqrt(3) / 2
-    y_padding = size
-
-    chart = chart.properties(
-        xlim=(-2 * x_padding, x_max + x_padding),
-        ylim=(-2 * y_padding, y_max + y_padding),
     )
 
     return chart
@@ -493,16 +506,24 @@ def _draw_network_grid(space, all_agent_data, agent_portrayal):
[…]
