+++ mesa/visualization/components/altair_components.py
@@ -336,12 +336,10 @@ def _draw_grid(space, agent_portrayal):
     """
     all_agent_data = get_agent_data(space, agent_portrayal)
 
-
     # Handle empty state
     if not all_agent_data:
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
 
-    
     match space:
         case Grid():
             return _draw_discrete_grid(space, all_agent_data, agent_portrayal)
@@ -392,8 +390,12 @@ def _draw_discrete_grid(space, all_agent_data, agent_portrayal):
         chart = chart.mark_point(size=30000 / length**2, filled=True)
 
     chart = chart.encode(
-        x=alt.X("x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))),
-        y=alt.Y("y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1)))
+        x=alt.X(
+            "x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))
+        ),
+        y=alt.Y(
+            "y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1))
+        ),
     )
 
     return chart
@@ -434,11 +436,15 @@ def _draw_legacy_grid(space, all_agent_data, agent_portrayal):
         chart = chart.mark_point(size=30000 / length**2, filled=True)
 
     chart = chart.encode(
-        x=alt.X("x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))),
-        y=alt.Y("y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1)))
+        x=alt.X(
+            "x", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.width - 1))
+        ),
+        y=alt.Y(
+            "y", axis=None, type=x_y_type, scale=alt.Scale(domain=(0, space.height - 1))
+        ),
     )
 
-    return chart 
+    return chart
 
 
 def _draw_hex_grid(space, all_agent_data, agent_portrayal):
@@ -470,12 +476,9 @@ def _draw_hex_grid(space, all_agent_data, agent_portrayal):
     x_scale = alt.Scale(domain=(-2 * x_padding, x_max + x_padding))
     y_scale = alt.Scale(domain=(-2 * y_padding, y_max + y_padding))
 
-
-
-
     encoding_dict = {
-        "x": alt.X("x", axis=None, type=x_y_type,scale=x_scale),
-        "y": alt.Y("y", axis=None, type=x_y_type,scale=y_scale),
+        "x": alt.X("x", axis=None, type=x_y_type, scale=x_scale),
+        "y": alt.Y("y", axis=None, type=x_y_type, scale=y_scale),
         "tooltip": [
             alt.Tooltip(key, type=alt.utils.infer_vegalite_type_for_pandas([value]))
             for key, value in all_agent_data[0].items()
@@ -510,14 +513,14 @@ def _draw_network_grid(space, all_agent_data, agent_portrayal):
     positions = [space.G.nodes[node].get("pos", (0, 0)) for node in space.G.nodes()]
     x_values = [p[0] for p in positions]
     y_values = [p[1] for p in positions]
-    
+
     # Add padding to the bounds
     padding = 0.1  # 10% padding
     x_min, x_max = min(x_values), max(x_values)
     y_min, y_max = min(y_values), max(y_values)
     x_range = x_max - x_min
     y_range = y_max - y_min
-    
+
     x_scale = alt.Scale(domain=(x_min - padding * x_range, x_max + padding * x_range))
[…]
