### 2025-02-17 f9ed4c1
paper: Fix - l109 I would suggest to move the Applications section to the start of the paper
