### 2025-02-18 dce72bf
updated boid to include direction
