### 2025-02-18 329bd24
removed generic_data_collection function
