+++ mesa/visualization/components/altair_components.py
@@ -249,73 +249,6 @@ def axial_to_pixel(q, r, size=1):
     return x, y
 
 
-def get_agent_data(space, agent_portrayal):
-    """Generic method to extract agent data for visualization across all space types.
-
-    Args:
-        space: Mesa space object
-        agent_portrayal: Function defining agent visualization properties
-
-    Returns:
-        List of agent data dictionaries with coordinates
-    """
-    all_agent_data = []
-
-    match space:
-        case Grid():
-            # New DiscreteSpace or experimental cell space
-            for cell in space.all_cells:
-                for agent in cell.agents:
-                    data = agent_portrayal(agent)
-                    data.update({"x": cell.coordinate[0], "y": cell.coordinate[1]})
-                    all_agent_data.append(data)
-
-        case _Grid():
-            # Legacy Grid
-            for content, (x, y) in space.coord_iter():
-                if not content:
-                    continue
-                agents = [content] if not hasattr(content, "__iter__") else content
-                for agent in agents:
-                    data = agent_portrayal(agent)
-                    data.update({"x": x, "y": y})
-                    all_agent_data.append(data)
-
-        case HexGrid():
-            # Hex-based grid
-            for content, (q, r) in space.coord_iter():
-                if content:
-                    for agent in content:
-                        data = agent_portrayal(agent)
-                        data.update({"q": q, "r": r})  # Store axial coordinates
-                        all_agent_data.append(data)
-
-        case NetworkGrid():
-            # Network grid
-            for node in space.G.nodes():
-                agents = space.G.nodes[node].get("agent", [])
-                if not isinstance(agents, list):
-                    agents = [agents] if agents else []
-
-                for agent in agents:
-                    if agent:
-                        data = agent_portrayal(agent)
-                        data.update({"node": node})  # Store node information
-                        all_agent_data.append(data)
-
-        case ContinuousSpace() | mesa.experimental.continuous_space.ContinuousSpace():
-            # Continuous space
-            for agent in space.agents:
-                data = agent_portrayal(agent)
-                data.update({"x": agent.pos[0], "y": agent.pos[1]})
-                all_agent_data.append(data)
-
-        case _:
-            raise NotImplementedError(f"Unsupported space type: {type(space)}")
-
-    return all_agent_data
-
-
 def _draw_grid(space, agent_portrayal):
     """Create Altair visualization for any supported space type.
 
@@ -334,29 +267,36 @@ def _draw_grid(space, agent_portrayal):
         and its agents.  Returns a text chart "No agents" if there are no agents.
 
     """
-    all_agent_data = get_agent_data(space, agent_portrayal)
-
-    # Handle empty state
-    if not all_agent_data:
+    # Handle empty state first
+    if not space.agents:
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
-
+    
     match space:
         case Grid():
-            return _draw_discrete_grid(space, all_agent_data)
+            return _draw_discrete_grid(space, agent_portrayal)
         case _Grid():
-            return _draw_legacy_grid(space, all_agent_data)
+            return _draw_legacy_grid(space, agent_portrayal)
         case HexGrid():
[…]
