### 2025-02-21 1b262e6
feat: added command console
