+++ mesa/visualization/command_console.py
@@ -1,0 +1,397 @@
+"""A command console interface for interactive Python code execution in the browser.
+
+This module provides a set of classes and functions to create an interactive Python console
+that can be embedded in a web browser. It supports command history, multi-line code blocks,
+and special commands for console management.
+Classes:
+    ConsoleEntry: Stores command console entries with their outputs and states.
+    CaptureOutput: Context manager for capturing stdout and stderr output.
+    InteractiveConsole: Custom interactive Python console with output capturing.
+    ConsoleManager: Manages the execution of Python code and console state.
+Functions:
+    format_command_html: Formats console commands as HTML for display.
+    format_output_html: Formats console output as HTML for display.
+    CommandConsole: Solara component for the interactive console interface.
+    >>> from mesa.visualization.command_console import CommandConsole
+    >>> CommandConsole(model=my_model)
+
+Notes:
+    - The console supports multi-line code blocks with proper indentation
+    - Output is captured and displayed with appropriate formatting
+    - Error messages are displayed in red with distinct styling
+    - The console maintains a history of commands and their outputs
+"""
+
+import code
+import io
+import sys
+from collections.abc import Callable
+from dataclasses import dataclass
+
+import solara
+from solara.components.input import use_change
+
+
+@dataclass
+class ConsoleEntry:
+    """A class to store command console entries.
+
+    Attributes:
+        command (str): The command entered
+        output (str): The output of the command
+        is_error (bool): Whether the entry represents an error
+        is_continuation (bool): Whether the entry is a continuation of previous command
+    """
+    command: str
+    output: str = ""
+    is_error: bool = False
+    is_continuation: bool = False
+
+    def __repr__(self):
+        """Return a string representation of the ConsoleEntry."""
+        return f"ConsoleEntry({self.command}, {self.output}, {self.is_error}, {self.is_continuation})"
+
+
+class CaptureOutput:
+    """A context manager for capturing stdout and stderr output.
+
+    This class provides a way to capture output that would normally be printed
+    to stdout and stderr during the execution of code within its context.
+    """
+    def __init__(self):
+        """Initialize the CaptureOutput context manager with empty string buffers."""
+        self.stdout = io.StringIO()
+        self.stderr = io.StringIO()
+        self._old_stdout = None
+        self._old_stderr = None
+
+    def __enter__(self):
+        """Set up the context manager by redirecting stdout and stderr.
+
+        Returns:
+            self: The context manager instance
+        """
+        self._old_stdout = sys.stdout
+        self._old_stderr = sys.stderr
+        sys.stdout = self.stdout
+        sys.stderr = self.stderr
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Restore the original stdout and stderr when exiting the context."""
+        sys.stdout = self._old_stdout
+        sys.stderr = self._old_stderr
+
+    def get_output(self):
+        """Retrieve and clear the captured output.
+
+        Returns:
+            tuple: A pair of strings (stdout_output, stderr_output)
+        """
+        output = self.stdout.getvalue()
+        error = self.stderr.getvalue()
+        self.stdout.seek(0)
+        self.stdout.truncate(0)
+        self.stderr.seek(0)
+        self.stderr.truncate(0)
+        return output, error
+
+
+class InteractiveConsole(code.InteractiveConsole):
+    """A custom interactive Python console with output capturing capabilities.
+
+    This class extends code.InteractiveConsole to provide output capturing functionality
+    when executing Python code interactively.
+
+    Args:
+        locals_dict (dict, optional): Dictionary of local variables. Defaults to None.
+    """
+    def __init__(self, locals_dict=None):
+        """Initialize the InteractiveConsole with the provided locals dictionary."""
[…]
