### 2025-02-21 1fadd28
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
