### 2025-02-22 904b369
stop running when running.value is false
