### 2025-02-24 bb12771
added direction to all agents
