+++ mesa/visualization/components/altair_components.py
@@ -3,10 +3,11 @@
 import warnings
 
 import altair as alt
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import solara
-from matplotlib.colors import to_rgba
+from matplotlib.colors import to_rgb
 
 import mesa
 from mesa.discrete_space import DiscreteSpace, Grid
@@ -183,11 +184,10 @@ def _draw_grid(space, agent_portrayal, propertylayer_portrayal):
         encoding_dict["color"] = alt.Color(
             "color:N",
             scale=alt.Scale(domain=unique_colors, range=unique_colors),
-            legend=None,
         )
     has_size = "size" in all_agent_data[0]
     if has_size:
-        encoding_dict["size"] = alt.Size("size", type="quantitative", legend=None)
+        encoding_dict["size"] = alt.Size("size", type="quantitative")
 
     agent_chart = (
         alt.Chart(
@@ -195,36 +195,37 @@ def _draw_grid(space, agent_portrayal, propertylayer_portrayal):
         )
         .mark_point(filled=True)
         .properties(width=300, height=300)
-        # .configure_view(strokeOpacity=0)  # hide grid/chart lines
     )
-    # This is the default value for the marker size, which auto-scales
-    # according to the grid area.
+
+    # This is the default value for the marker size, which auto-scales according to the grid area.
     if not has_size:
         length = min(space.width, space.height)
-        chart = agent_chart.mark_point(size=30000 / length**2, filled=True)
+        agent_chart = agent_chart.mark_point(size=30000 / length**2, filled=True)
 
     if propertylayer_portrayal is not None:
-        base_width = agent_chart.properties().width
-        base_height = agent_chart.properties().height
+        chart_width = agent_chart.properties().width
+        chart_height = agent_chart.properties().height
         chart = chart_property_layers(
             space=space,
             propertylayer_portrayal=propertylayer_portrayal,
-            base_width=base_width,
-            base_height=base_height,
+            chart_width=chart_width,
+            chart_height=chart_height,
         )
+        chart = chart + agent_chart
+    else:
+        chart = agent_chart
 
-    chart = chart + agent_chart
     return chart
 
 
-def chart_property_layers(space, propertylayer_portrayal, base_width, base_height):
+def chart_property_layers(space, propertylayer_portrayal, chart_width, chart_height):
     """Creates Property Layers in the Altair Components.
 
     Args:
         space: the ContinuousSpace instance
         propertylayer_portrayal:Dictionary of PropertyLayer portrayal specifications
-        base_width: width of the agent chart to maintain consistency with the property charts
-        base_height: height of the agent chart to maintain consistency with the property charts
+        chart_width: width of the agent chart to maintain consistency with the property charts
+        chart_height: height of the agent chart to maintain consistency with the property charts
     Returns:
         Altair Chart
     """
@@ -245,7 +246,7 @@ def chart_property_layers(space, propertylayer_portrayal, base_width, base_heigh
 
         data = layer.data.astype(float) if layer.data.dtype == bool else layer.data
 
-        if (space.width, space.height) is not data.shape:
+        if (space.width, space.height) != data.shape:
             warnings.warn(
                 f"Layer {layer_name} dimensions ({data.shape}) do not match space dimensions ({space.width}, {space.height}).",
                 UserWarning,
@@ -265,30 +266,21 @@ def chart_property_layers(space, propertylayer_portrayal, base_width, base_heigh
             }
         )
 
-        # Add RGBA color if "color" is in portrayal
         if "color" in portrayal:
-            df["color"] = df["value"].apply(
-                lambda val,
-                portrayal=portrayal,
[…]
