### 2025-02-27 4d09230
vectorization and cache
