+++ mesa/examples/basic/boid_flockers/app.py
@@ -1,7 +1,6 @@
 import os
 import sys
 
-import numpy as np
 from matplotlib.markers import MarkerStyle
 
 sys.path.insert(0, os.path.abspath("../../../.."))
@@ -9,20 +8,26 @@ sys.path.insert(0, os.path.abspath("../../../.."))
 from mesa.examples.basic.boid_flockers.model import BoidFlockers
 from mesa.visualization import Slider, SolaraViz, make_space_component
 
+# Pre-compute markers for different angles (e.g., every 10 degrees)
+MARKER_CACHE = {}
+for angle in range(0, 360, 10):
+    marker = MarkerStyle(10)
+    marker._transform = marker.get_transform().rotate_deg(angle)
+    MARKER_CACHE[angle] = marker
+
 
 def boid_draw(agent):
     neighbors = len(agent.neighbors)
 
+    # Calculate the angle
+    deg = agent.angle
+    # Round to nearest 10 degrees
+    rounded_deg = round(deg / 10) * 10 % 360
+
     if neighbors <= 1:
-        deg = np.degrees(np.arctan2(agent.direction[0], agent.direction[1]))
-        marker = MarkerStyle(10)
-        marker._transform = marker.get_transform().rotate_deg(deg)
-        return {"color": "red", "size": 20, "marker": marker}
+        return {"color": "red", "size": 20, "marker": MARKER_CACHE[rounded_deg]}
     elif neighbors >= 2:
-        deg = np.degrees(np.arctan2(agent.direction[0], agent.direction[1]))
-        marker = MarkerStyle(10)
-        marker._transform = marker.get_transform().rotate_deg(deg)
-        return {"color": "green", "size": 20, "marker": marker}
+        return {"color": "green", "size": 20, "marker": MARKER_CACHE[rounded_deg]}
 
 
 model_params = {
