+++ mesa/visualization/solara_viz.py
@@ -58,8 +58,8 @@ def SolaraViz(
     simulator: Simulator | None = None,
     model_params=None,
     name: str | None = None,
-    additional_imports: dict | None = None,
     use_threads: bool = False,
+    additional_imports: dict | None = None,
 ):
     """Solara visualization component.
 
