+++ mesa/visualization/components/altair_components.py
@@ -2,8 +2,10 @@
 
 import contextlib
 import math
+import itertools
 import warnings
 from collections.abc import Callable
+from functools import lru_cache
 
 import solara
 
@@ -17,7 +19,7 @@ with contextlib.suppress(ImportError):
 from mesa.experimental.cell_space import Grid
 from mesa.space import ContinuousSpace, NetworkGrid, _Grid
 from mesa.visualization.utils import update_counter
-
+import numpy as np
 
 def make_space_altair(*args, **kwargs):
     """Create an Altair chart component for visualizing model space (deprecated).
@@ -66,7 +68,6 @@ def make_altair_space(
         function: A function that creates a SpaceAltair component
     """
     if agent_portrayal is None:
-
         def agent_portrayal(a):
             return {"id": a.unique_id}
 
@@ -226,20 +227,19 @@ def SpaceAltair(
         A solara.FigureAltair instance, which is a Solara component that
         renders the Altair chart.
 
-
     """
+    # Force update on dependencies change
     update_counter.get()
     space = getattr(model, "grid", None)
     if space is None:
-        # Sometimes the space is defined as model.space instead of model.grid
         space = model.space
 
     chart = _draw_grid(space, agent_portrayal)
-    # Apply post-processing if provided
     if post_process is not None:
         chart = post_process(chart)
 
-    solara.FigureAltair(chart)
+    # Return the rendered chart
+    return solara.FigureAltair(chart)
 
 
 def axial_to_pixel(q, r, size=1):
@@ -272,12 +272,12 @@ def _draw_grid(space, agent_portrayal):
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
 
     match space:
+        case HexGrid():
+            return _draw_hex_grid(space, agent_portrayal)
         case Grid():
             return _draw_discrete_grid(space, agent_portrayal)
         case _Grid():
             return _draw_legacy_grid(space, agent_portrayal)
-        case HexGrid():
-            return _draw_hex_grid(space, agent_portrayal)
         case NetworkGrid():
             return _draw_network_grid(space, agent_portrayal)
         case ContinuousSpace() | mesa.experimental.continuous_space.ContinuousSpace():
@@ -289,55 +289,53 @@ def _draw_grid(space, agent_portrayal):
 def _draw_discrete_grid(space, agent_portrayal):
     """Create Altair visualization for Discrete Grid."""
     all_agent_data = []
+    
+    # Collect agent data
     for cell in space.all_cells:
         for agent in cell.agents:
             data = agent_portrayal(agent)
-            data.update({"x": cell.coordinate[0], "y": cell.coordinate[1]})
+            data.update({
+                "x": float(cell.coordinate[0]), 
+                "y": float(cell.coordinate[1])
+            })
             all_agent_data.append(data)
 
     if not all_agent_data:
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
 
-    invalid_tooltips = ["color", "size", "x", "y"]
-    x_y_type = "ordinal"
+    # Create base chart
+    base = alt.Chart(alt.Data(values=all_agent_data)).properties(
+        width=280, height=280
+    )
 
-    encoding_dict = {
-        "x": alt.X("x", axis=None, type=x_y_type),
-        "y": alt.Y("y", axis=None, type=x_y_type),
-        "tooltip": [
-            alt.Tooltip(key, type=alt.utils.infer_vegalite_type_for_pandas([value]))
[…]
