+++ mesa/visualization/components/altair_components.py
@@ -1,8 +1,8 @@
 """Altair based solara components for visualization mesa spaces."""
 
 import contextlib
-import math
 import itertools
+import math
 import warnings
 from collections.abc import Callable
 from functools import lru_cache
@@ -16,10 +16,12 @@ from mesa.experimental.cell_space import HexGrid
 with contextlib.suppress(ImportError):
     import altair as alt
 
+import numpy as np
+
 from mesa.experimental.cell_space import Grid
 from mesa.space import ContinuousSpace, NetworkGrid, _Grid
 from mesa.visualization.utils import update_counter
-import numpy as np
+
 
 def make_space_altair(*args, **kwargs):
     """Create an Altair chart component for visualizing model space (deprecated).
@@ -68,6 +70,7 @@ def make_altair_space(
         function: A function that creates a SpaceAltair component
     """
     if agent_portrayal is None:
+
         def agent_portrayal(a):
             return {"id": a.unique_id}
 
@@ -289,36 +292,33 @@ def _draw_grid(space, agent_portrayal):
 def _draw_discrete_grid(space, agent_portrayal):
     """Create Altair visualization for Discrete Grid."""
     all_agent_data = []
-    
+
     # Collect agent data
     for cell in space.all_cells:
         for agent in cell.agents:
             data = agent_portrayal(agent)
-            data.update({
-                "x": float(cell.coordinate[0]), 
-                "y": float(cell.coordinate[1])
-            })
+            data.update(
+                {"x": float(cell.coordinate[0]), "y": float(cell.coordinate[1])}
+            )
             all_agent_data.append(data)
 
     if not all_agent_data:
         return alt.Chart().mark_text(text="No agents").properties(width=280, height=280)
 
     # Create base chart
-    base = alt.Chart(alt.Data(values=all_agent_data)).properties(
-        width=280, height=280
-    )
+    base = alt.Chart(alt.Data(values=all_agent_data)).properties(width=280, height=280)
 
     # Configure encodings
     encodings = {
         "x": alt.X(
             "x:Q",
-            scale=alt.Scale(domain=[0, space.width-1]),
-            axis=alt.Axis(grid=True)  # Enable grid
+            scale=alt.Scale(domain=[0, space.width - 1]),
+            axis=alt.Axis(grid=True),  # Enable grid
         ),
         "y": alt.Y(
-            "y:Q", 
-            scale=alt.Scale(domain=[0, space.height-1]),
-            axis=alt.Axis(grid=True)  # Enable grid
+            "y:Q",
+            scale=alt.Scale(domain=[0, space.height - 1]),
+            axis=alt.Axis(grid=True),  # Enable grid
         ),
     }
 
@@ -326,12 +326,12 @@ def _draw_discrete_grid(space, agent_portrayal):
     if "color" in all_agent_data[0]:
         encodings["color"] = alt.Color("color:N")
 
-    # Add size encoding if present 
+    # Add size encoding if present
     if "size" in all_agent_data:
         encodings["size"] = alt.Size("size:Q")
     else:
         # Default size based on grid dimensions
-        point_size = 30000 / min(space.width, space.height)**2
+        point_size = 30000 / min(space.width, space.height) ** 2
         base = base.mark_point(size=point_size, filled=True)
 
     # Create final chart with encodings
@@ -389,10 +389,16 @@ def _draw_legacy_grid(space, agent_portrayal):
 
     chart = chart.encode(
         x=alt.X(
[…]
