+++ mesa/visualization/solara_viz.py
@@ -86,8 +86,13 @@ def SolaraViz(
         simulator: A simulator that controls the model (optional)
         model_params (dict, optional): Parameters for (re-)instantiating a model.
             Can include user-adjustable parameters and fixed parameters. Defaults to None.
-        name (str | None, optional): Name of the visualization. Defaults to the models class name.
-        additional_imports (dict, optional): Dictionary of names to objects to import into the command console.
+        name (str | None, optional): Name of the visualization. Defaults to the model's class name.
+        **console_kwargs (dict, optional): Arguments to pass to the command console.
+            Currently supported arguments:
+            - additional_imports: Dictionary of names to objects to import into the command console.
+                - Example:
+                    >>> console_kwargs = {"additional_imports": {"numpy": np}}
+                    >>> SolaraViz(model, console_kwargs=console_kwargs)
 
     Returns:
         solara.component: A Solara component that renders the visualization interface for the model.
@@ -103,7 +108,7 @@ def SolaraViz(
         - The `play_interval` argument controls the speed of the model's automatic stepping. A lower
           value results in faster stepping, while a higher value results in slower stepping.
         - The `render_interval` argument determines how often plots are updated during simulation. Higher values
-          reduce update frequency,resulting in faster execution.
+          reduce update frequency, resulting in faster execution.
     """
     if components == "default":
         components = [
