+++ mesa/visualization/__init__.py
@@ -9,6 +9,7 @@ from mesa.visualization.mpl_space_drawing import (
     draw_space,
 )
 
+from .command_console import CommandConsole
 from .components import make_plot_component, make_space_component
 from .components.altair_components import make_space_altair
 from .solara_viz import JupyterViz, SolaraViz
@@ -22,4 +23,5 @@ __all__ = [
     "make_plot_component",
     "make_space_altair",
     "make_space_component",
+    "CommandConsole",
 ]
