+++ mesa/visualization/command_console.py
@@ -189,7 +189,9 @@ class ConsoleManager:
                     f"Command: {entry.command}, \nOutput: {entry.output if entry.output else None}\n"
                 )
                 for entry in self.history
-                if entry.command != "[history]" and entry.command != "[tips]" and entry.command != ""
+                if entry.command != "[history]"
+                and entry.command != "[tips]"
+                and entry.command != ""
             ]
 
             self.history.append(
