### 2025-03-02 09f7758
fix: auto scrolling
