+++ mesa/visualization/components/altair_components.py
@@ -3,7 +3,6 @@
 import warnings
 
 import altair as alt
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import solara
@@ -196,6 +195,7 @@ def _draw_grid(space, agent_portrayal, propertylayer_portrayal):
         .mark_point(filled=True)
         .properties(width=300, height=300)
     )
+    base_chart = None
 
     # This is the default value for the marker size, which auto-scales according to the grid area.
     if not has_size:
@@ -205,20 +205,21 @@ def _draw_grid(space, agent_portrayal, propertylayer_portrayal):
     if propertylayer_portrayal is not None:
         chart_width = agent_chart.properties().width
         chart_height = agent_chart.properties().height
-        chart = chart_property_layers(
+        base_chart = chart_property_layers(
             space=space,
             propertylayer_portrayal=propertylayer_portrayal,
             chart_width=chart_width,
             chart_height=chart_height,
+            agent_chart=agent_chart,
         )
-        chart = chart + agent_chart
     else:
-        chart = agent_chart
-    chart = chart + agent_chart
-    return chart
+        base_chart = agent_chart
+    return base_chart
 
 
-def chart_property_layers(space, propertylayer_portrayal, chart_width, chart_height):
+def chart_property_layers(
+    space, propertylayer_portrayal, chart_width, chart_height, agent_chart
+):
     """Creates Property Layers in the Altair Components.
 
     Args:
@@ -226,6 +227,7 @@ def chart_property_layers(space, propertylayer_portrayal, chart_width, chart_hei
         propertylayer_portrayal:Dictionary of PropertyLayer portrayal specifications
         chart_width: width of the agent chart to maintain consistency with the property charts
         chart_height: height of the agent chart to maintain consistency with the property charts
+        agent_chart: the agent chart to layer with the property layers on the grid
     Returns:
         Altair Chart
     """
@@ -235,7 +237,7 @@ def chart_property_layers(space, propertylayer_portrayal, chart_width, chart_hei
     except AttributeError:
         # new style spaces
         property_layers = space._mesa_property_layers
-    base = None
+    base = agent_chart
     for layer_name, portrayal in propertylayer_portrayal.items():
         layer = property_layers.get(layer_name, None)
         if not isinstance(
@@ -267,21 +269,38 @@ def chart_property_layers(space, propertylayer_portrayal, chart_width, chart_hei
         )
 
         if "color" in portrayal:
-            # any value less than vmin will be mapped to the color corresponding to vmin
-            # any value more than vmax will be mapped to the color corresponding to vmax
+            # Create a function to map values to RGBA colors with proper opacity scaling
             def apply_rgba(val, vmin=vmin, vmax=vmax, alpha=alpha, portrayal=portrayal):
-                a = (val - vmin) / (vmax - vmin)
-                a = max(0, min(a, 1))  # to ensure that a is between 0 and 1
-                a *= alpha  # vmax will have an opacity corresponding to alpha
+                """Maps data values to RGBA colors with opacity based on value magnitude.
+
+                Args:
+                    val: The data value to convert
+                    vmin: The smallest value for which the color is displayed in the colorbar
+                    vmax: The largest value for which the color is displayed in the colorbar
+                    alpha: The opacity of the color
+                    portrayal: The specifics of the current property layer in the iterative loop
+
+                Returns:
+                    String representation of RGBA color
+                """
+                # Normalize value to range [0,1] and clamp
+                normalized = max(0, min((val - vmin) / (vmax - vmin), 1))
+
+                # Scale opacity by alpha parameter
+                opacity = normalized * alpha
+
+                # Convert color to RGB components
[…]
