### 2025-03-05 #7668 @martibosch
Hello!

sorry for the delay. I started the review yesterday and I see that several of my remarks on my manuscript have already been pointed out by @jofmi, so I re-generated the pdf.

## comments on the paper

- I completely agree with @jofmi on the comments on the "Statement of need" section. I find the new version much improved but to me it still feels a bit like a historical recap - I believe it could be more in line with the JOSS guidelines.
- I find that we miss a review of related tools. Besides agentpy, a quick github search of "agent based modeling python" pointed me to [Melodie](https://github.com/ABM4ALL/Melodie) (which also has a JOSS paper), [helipad](https://github.com/charwick/helipad), [pythonabm](https://github.com/JackToppen/pythonabm), [bptk_py](https://github.com/transentis/bptk_py) (a bit more system dynamics focused but still sort of related). Of course many results are intrasnscendent and unmantained repositories, but at first glance the ones that I listed merit consideration and should potentially be referenced in this paper.
- In the section "Agent management", the `AgentSet` class is mentioned but it does not appear explicitly in the code. I infer that it is instantiated automatically with the models and then made available through the `agents` attribute, but I think it is a good idea to clarify it. This also applies [to the introductory tutorial](https://mesa.readthedocs.io/stable/getting_started.html#agentset-functionality).

### typos and minor syntactic comments
- line 21: there is a missing space between "NetLogo" and "in"
- line 34: I find that the word "everything" appears too informal in this context, I would change it to something like "has been applied to modeling a wide range of phenomena from economics ..."
- line 126: broken SolaraViz link
