### 2025-03-06 #7668 @martibosch
Hello,

overall everything in the code seems very solid and the user interface is very clear so I do not have any major remark regarding that. This is obviously a very valuable library and I strongly appreciate that is natively integrated within the scientific Python ecosystem.

 My comments are thus on the documentation and user guide, except for the last two which are a bit related to the code and packaging structure.

### comments on the documentation and user guide

- Like @jofmi, I find that the introductory and visualization tutorials would be better served as self-contained notebooks with a "pip install" cell to install the tutorial's requirements. For instance, the introductory tutorial only requires `pip install mesa` whereas the visualization tutorial requires `pip install "mesa[viz]"`. This may seem obvious but not if someone jumpstarts to the tutorials without going to "Installation Options" section on the documentation index.
- Note again the comment on the paper regarding the `AgentSet` class and the `agents` attribute and how it applies to the introductory tutorial as well.
- In line with the point above (and again @jofmi), I also had a hard time knowing which `MoneyModel` I had to use in the visualization tutorial. Many `MoneyModel` classes are defined in the introductory tutorial, so at the end I had to go to https://github.com/projectmesa/mesa/blob/main/docs/tutorials/MoneyModel.py to get the one that worked. Although repetitve, I think it would be better to include the class definition in the visualization tutorial.
- It may be a good idea to explicitly mention the `number_processes` argument of `mesa.batch_run` in the getting started page. The first thing I thought was "this is probably embarrassingly parallel in most cases so it would be cool to run it with dask or the like", but when I went to the API documentation I saw that this option already exists.
- Following from the point above, why is it "strongly advised to only run in parallel using a normal python file (so don’t try to do it in a jupyter notebook)"? It would be good if users could run in parallel in notebooks too. I am not an expert on multiprocessing in Python, but in my experience I had no issues runing embarrassingly parallel code through Jupyter ([I actually used dask](https://github.com/martibosch/pylandstats/blob/main/src/pylandstats/multilandscape.py#L157-L166), but I do not see why it should be different with multiprocessing).
- I find confusing that in the "Mesa core examples" section you refer to the mesa-examples repository but then the code snippets in the showcased examples correspond to files in https://github.com/projectmesa/mesa/tree/main/mesa/examples.  It may be better to put everything in one place and mention in the documentation that users can clone it and go to each example's folder instead of having them copy each snippet in a separate file. In any case, are the basic examples packaged as a core mesa module for a particular reason?
- Still related to the examples: the source distribution is quite big (2.4 MB), is that intended? In most of the Python libraries that I use the docs are excluded from the source distribution (in your case, when uncompressed the docs account to 2.5 out of the total 3.4 MB), is there any reason why this is different with mesa?
