### 2025-03-06 #7668 @EwoutH
@martibosch Thanks a lot for your extensive review!

I just started my new job this Monday, so I probably will take some time to address your comments. I'll also check if my co-maintainers/authors want to pick anything up.

We might open some PRs to the Project Mesa repository, and will include you as reviewer where relevant to the paper.
