### 2025-03-06 projectmesa/mesa#2666 @EwoutH
@projectmesa/maintainers I updated the first post in this issue with the new comments from Martí Bosch, our second reviewer. Those have the prefix **[from Marti]**.

Now, let's try to distribute some task. Please mark your checkbox if you're okay with picking this up, and if not or if you have an other idea or thing you want to do, let us know.
- I will pick up the **documentation comments**
- @jackiekazil can you pick up the remaining comments on **the paper itself**?
- @tpike3 are you willing to pick up the comments on the **introduction tutorial**, the **visualisation tutorial**, or **both**?
- @wang-boyu are you willing to pick up the **typing comments** or **one of the tutorials**?

@quaquel I know you are very busy right now, but maybe you could review some of the incoming PRs on the paper.

@martibosch, I temporarily [invited you](https://github.com/projectmesa/mesa/invitations) to become a Triager in Mesa, please consider accepting so you can also review these PRs. This will only be for the duration of the paper review.

@Sahil-Chhoker if you want to contribute anything to our paper as collaborator, I'm good with that. We can mention you in the acknowledgements, if that matters to you.

Let's drive this home!
