### 2025-03-11 projectmesa/mesa#2666 @jackiekazil
The line numbers are quite helpful in the comments, but with all the edits -- I am struggling to track down which line. Does anyone know how to refer to a previous version of the paper?

Example of what I am struggling to identify specifically (but this probably won't be the only one) --
l41 For consistency with the other sections, please add a code example of instantiating a model and passing a seed
l46 Similarly, please add a code example of instantiating and subclassing an agent
