### 2025-03-11 projectmesa/mesa#2666 @EwoutH
@jackiekazil would you want to go back in de PDF or Markdown version?

- PDF: https://github.com/openjournals/joss-reviews/issues/7668#issuecomment-2591397878 (Jan 15) / https://github.com/openjournals/joss-reviews/issues/7668#issuecomment-2700862023 (recent)
- Markdown: Check the [git history](https://github.com/projectmesa/mesa/commits/paper_2024/paper/paper.md)

Thanks for the PR!
