### 2025-03-12 projectmesa/mesa#2666 @EwoutH
No it's just a version from git history I'm afraid.

Our CI does build a PDF on every chance, so for every commit there should be a PDF available in [our CI here](https://github.com/projectmesa/mesa/actions/workflows/paper.yml).
