### 2025-03-13 projectmesa/mesa#2716 @EwoutH
@mukulpythondev Awesome! I assigned you, feel free to discuss a potential solution here before opening a PR if you want.
