### 2025-03-13 projectmesa/mesa#2716 @adamamer20
I remember running into a similar problem in `mesa-frames`

https://github.com/projectmesa/mesa-frames/discussions/12

I think using generics might work. The only limit I found was that generics do not work well with inheritance but this shouldn't be an issue in this case.
