### 2025-03-19 #7668 @jofmi
Hi, I looked at the new version and it looks good! :)

Only one point remaining from my side: i would remove line 110-113 if this comment is true and the feature is already in a ready-to-merge PR and simply wait for this PR to be merged before the release of this paper https://github.com/projectmesa/mesa/pull/2675#issuecomment-2647649341

Oh, and one more typo: missing space in line 63
