### 2025-03-19 #7668 @EwoutH
@jofmi thanks, good catch! Both are fixed both in https://github.com/projectmesa/mesa/pull/2730.
