### 2025-03-21 #7668 @martibosch
Hello,

overall I feel like all of my comments have been adressed except the review of related tools (e.g., in the statement of need):

> * I find that we miss a review of related tools. Besides agentpy, a quick github search of "agent based modeling python" pointed me to [Melodie](https://github.com/ABM4ALL/Melodie) (which also has a JOSS paper), [helipad](https://github.com/charwick/helipad), [pythonabm](https://github.com/JackToppen/pythonabm), [bptk_py](https://github.com/transentis/bptk_py) (a bit more system dynamics focused but still sort of related). Of course many results are intrasnscendent and unmantained repositories, but at first glance the ones that I listed merit consideration and should potentially be referenced in this paper.

Maybe there is a good reason why these other tools are not mentioned, in which case, may you briefly share it here?

After that I will just quickly go again through the manuscript and that should be it!
