### 2025-03-21 #7668 @EwoutH
Thanks for getting back! On the other tools, I noted earlier:

> There were some comments on the Statement of Need and comparison to other tools. Personally I think @jackiekazil improved it quite well and it now captures both on a high-level. JOSS isn't meant to do deep-dive comparisons with other tools as far as I know. And our paper is already on the long side.

JOSS papers are meant to be short and concise, not deep-dives into all possible alternatives and how they compare. I think we found a nice balance by mentioning some in the Statement of Need.
