### 2025-03-21 #7668 @martibosch
oh, sorry I missed that bit. In such a case I leave it up to @gkthiruvathukal and then check the two last items from the checklist.

I just went through the paper again and would just like to raise again this:

> * In the section "Agent management", the `AgentSet` class is mentioned but it does not appear explicitly in the code. I infer that it is instantiated automatically with the models and then made available through the `agents` attribute, but I think it is a good idea to clarify it. This also applies [to the introductory tutorial](https://mesa.readthedocs.io/stable/getting_started.html#agentset-functionality).

I think it woud be appropriate to explicitly mention that the `AgentSet` class is accessed through the `agents` attribute (lines 88-90). But I leave it up to the authors.

Typos:

line 63: missing space between ")" and "to"
