### 2025-03-21 #7668 @EwoutH
Thanks again for your critical look. Really good feedback on `model.agents` AgentSet, we should clarify that. I've done so in https://github.com/projectmesa/mesa/pull/2733, by adding this line:

> When agents are created, they automatically register with the model via `model.register_agent(self)`. This registration automatically adds the agent to an `AgentSet` that's accessible through the model's `agents` property. Additional `AgentSet` instances for each agent type are maintained and available through `model.agents_by_type`. These collections are automatically updated when agents are added or removed from the model.

Does that explain it properly?

Missing space was fixed in https://github.com/projectmesa/mesa/pull/2730.
