### 2025-03-21 #7668 @EwoutH
@gkthiruvathukal I was looking further in the "State of the field" requirement, and I got a bit confused. It's mentioned in the reviewer checklist, but nowhere in the [JOSS Paper Format](https://joss.readthedocs.io/en/latest/paper.html). Do we need to cover this more extensively then we currently have in the Statement of Need, and if so, where should we do this?
