### 2025-03-21 #7668 @gkthiruvathukal
@EwoutH  I believe this pertains to related work (i.e. others doing similar work). Good to see that this submission is getting closer. (Thanks @martibosch for your continued review comments.)

I'm going to loop in @danielskatz to see whether he has any additional input as this topic would seemingly be relevant to the written part of a JOSS submission (the paper format).
