### 2025-03-21 #7668 @danielskatz
The state of the field is supposed to make clear how the software in this paper relates to other software. It can be in the Statement of Need section or elsewhere. It's up to the reviewers to decide if it's sufficient.
