### 2025-03-22 #7668 @EwoutH
Thanks for both your insights!

@jofmi and @martibosch considering this, could you both let me know if you find the current discussion in the Statement of Need satisfactory or if you would like to suggest changes?
