### 2025-03-24 projectmesa/mesa#2716 @EwoutH
Thanks for your work, it's a difficult issue and glad you learned something for it.

If you have any further insights that might help a future contributor help understand this issue (or know what not to try), please share them!
