### 2025-03-24 projectmesa/mesa#2716 @mukulpythondev
Thank you! I really appreciate the opportunity to work on this. The main challenge was ensuring proper type inference in AgentSet and maintaining subclass types in model.agents. Future contributors might want to explore alternative ways to enforce type safety in collections while keeping backward compatibility. Happy to discuss further if needed!
