### 2025-03-25 #7668 @gkthiruvathukal
@EwoutH Can you please fix the above (invalid DOIs)?
