### 2025-03-26 #7668 @EwoutH
- Authors and ORCIDs are up to date
- Release with latest changes from review is created: [v3.1.5](https://github.com/projectmesa/mesa/releases/tag/v3.1.5)
- Archive + DOI is created: `10.5281/zenodo.15089114` [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15089114.svg)](https://doi.org/10.5281/zenodo.15089114)
- Title, authors and license match
