### 2025-03-27 #7668 @EwoutH
> I find this submission to be really well done. (This reminded me of the old days working with NetLogo!)

Thanks, really appreciated! I also started my ABM adventure with NetLogo, I think it was early 2018. Still an amazing piece of software by an amazing team!

> One quick thing. The paper says version 3.14 but the current tagged release is 3.15. This might not be a big deal but if you'd like to have a chance to correct this before I recommend accept, let me know.

Good catch! I updated it.
