### 2025-03-27 #7668 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
