### 2025-03-27 #7668 @danielskatz
@EwoutH - I think the paper looks mostly good, with a few small things to fix in https://github.com/projectmesa/mesa/pull/2738. Have you confirmed that the proof is ok with you otherwise?  Also, I note there are no funding statements in the acknowledgments - should there be?
