### 2025-03-27 #7668 @EwoutH
@quaquel @corvince @tpike3 @wang-boyu @rht @jackiekazil please check the final proof of the paper one last time! https://raw.githubusercontent.com/openjournals/joss-papers/joss.07668/joss.07668/10.21105.joss.07668.pdf
