### 2025-03-27 #7668 @EwoutH
Thanks for looking at it!

> Have you confirmed that the proof is ok with you otherwise?

Yes, I’m good. Let’s give my co-authors one last opportunity to double-check it. 24 hours should be enough.

> Also, I note there are no funding statements in the acknowledgments - should there be?

That’s correct. We’re volunteer based (somewhat unfortunately).
