### 2025-03-28 #7668 @danielskatz
just a final check of the changes
