### 2025-03-28 #7668 @danielskatz
Congratulations to  @EwoutH (Ewout ter Hoeven) and co-authors on your publication!!

The DOI isn't resolving for me, so I will keep this issue open until it does.
