### 2025-03-28 #7668 @danielskatz
The DOI now works for me!
