### 2025-05-14 #7668 @ritika-giri
Absolutely - the USRSE'25 is a conference for anyone at the intersection of code and research, with a specific focus on open science. This year's theme is **Code, Practices, and People** and [here](https://us-rse.org/usrse25/participate/) you can see what types of submissions we invite, and the formats we accept. The idea is to form a community of research software professionals (loosely defined) since these are the people who make research software happen, but often don't have enough visibility or formalized support from institutions yet.

As a note, since this is a published paper - you can submit to any of the tracks (talk, poster, workshop, notebook) other than the paper track. To get a better idea of the audience, feel free to peruse [last year's program](https://us-rse.org/usrse24/program/).

Also feel free to [join](https://us-rse.org/join/) The United States Research Software Engineer Association, and connect with folks over Slack. Let me know if there is anything in particular you'd like to hear about!
