### 2025-05-14 #7668 @cmaimone
Unfortunately we don't currently have any travel support.  If we do, it will be limited to those in the US (we may have missed that you weren't in the US if someone reached out).

For a workshop, there can be multiple presenters/instructors if that's what you're asking.
