### 2025-05-14 #7668 @EwoutH
That's unfortunate to hear, but thanks for the information.

@jackiekazil, @tpike3 and @boyu I don't think I can convince my work to both attend this conference and fund travel. Since you're already in the US, is either of you interested in giving a talk or hosting a workshop? (details [here](https://us-rse.org/usrse25/participate/))
